@@ -56,7 +56,7 @@ namespace CSRedis.Internal.Commands
 
             protected virtual T[] Read(long count, RedisReader reader)
             {
-				if (count < 0) return null;
+                if (count < 0) return null;
                 T[] array = new T[count];
                 for (int i = 0; i < array.Length; i++)
                     array[i] = _memberCommand.Parse(reader);
@@ -80,8 +80,8 @@ namespace CSRedis.Internal.Commands
             {
                 reader.ExpectType(RedisMessage.MultiBulk);
                 long count = reader.ReadInt(false);
-				if (count == -1) return default(T); //使用 BLPop 命令在 RedisArray.cs 中报错的解决办法。 #22
-				T[] array = new T[count];
+                if (count == -1) return default(T); //使用 BLPop 命令在 RedisArray.cs 中报错的解决办法。 #22
+                T[] array = new T[count];
                 for (int i = 0; i < array.Length; i++)
                     array[i] = _command.Parse(reader);
                 return array[_index];
@@ -94,12 +94,13 @@ namespace CSRedis.Internal.Commands
                 : base(new RedisString(command, args))
             { }
         }
-		public class Bytes : Generic<byte[]> {
-			public Bytes(string command, params object[] args)
-				: base(new RedisBytes(command, args)) { }
-		}
+        public class Bytes : Generic<byte[]>
+        {
+            public Bytes(string command, params object[] args)
+                : base(new RedisBytes(command, args)) { }
+        }
 
-		public class StrongPairs<T1, T2> : Generic<Tuple<T1, T2>>
+        public class StrongPairs<T1, T2> : Generic<Tuple<T1, T2>>
         {
             public StrongPairs(RedisCommand<T1> command1, RedisCommand<T2> command2, string command, params object[] args)
                 : base(new RedisTuple.Generic<T1, T2>.Repeating(command1, command2, command, args))
@@ -120,5 +121,5 @@ namespace CSRedis.Internal.Commands
                 : base(new RedisString.Converter<T1>(null), new RedisString.Converter<T2>(null), command, args)
             { }
         }
-	}
+    }
 }
