@@ -4,29 +4,29 @@ using System.Globalization;
 
 namespace CSRedis.Internal.Commands
 {
-    class RedisFloat : RedisCommand<double>
+    class RedisFloat : RedisCommand<decimal>
     {
         public RedisFloat(string command, params object[] args)
             : base(command, args)
         { }
 
-        public override double Parse(RedisReader reader)
+        public override decimal Parse(RedisReader reader)
         {
             return FromString(reader.ReadBulkString());
         }
 
-        static double FromString(string input)
+        static decimal FromString(string input)
         {
-            return Double.Parse(input, NumberStyles.Float, CultureInfo.InvariantCulture);
+            return decimal.Parse(input, NumberStyles.Any);
         }
 
-        public class Nullable : RedisCommand<double?>
+        public class Nullable : RedisCommand<decimal?>
         {
             public Nullable(string command, params object[] args)
                 : base(command, args)
             { }
 
-            public override double? Parse(RedisReader reader)
+            public override decimal? Parse(RedisReader reader)
             {
                 string result = reader.ReadBulkString();
                 if (string.IsNullOrEmpty(result))
