@@ -4,7 +4,7 @@ using System.ComponentModel;
 
 namespace CSRedis.Internal.Commands
 {
-	class RedisString : RedisCommand<string>
+    class RedisString : RedisCommand<string>
     {
         public RedisString(string command, params object[] args)
             : base(command, args)
