@@ -3,7 +3,9 @@ using CSRedis.Internal.IO;
 using CSRedis.Internal.Utilities;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 
@@ -35,6 +37,14 @@ namespace CSRedis
         #endregion
 
         #region Keys
+        public static RedisInt Touch(params string[] keys)
+        {
+            return new RedisInt("TOUCH", keys);
+        }
+        public static RedisInt UnLink(params string[] keys)
+        {
+            return new RedisInt("UNLINK", keys);
+        }
         public static RedisInt Del(params string[] keys)
         {
             return new RedisInt("DEL", keys);
@@ -46,6 +56,10 @@ namespace CSRedis
         public static RedisBool Exists(string key)
         {
             return new RedisBool("EXISTS", key);
+        }
+        public static RedisInt Exists(string[] keys)
+        {
+            return new RedisInt("EXISTS", keys);
         }
         public static RedisBool Expire(string key, TimeSpan expiration)
         {
@@ -143,7 +157,7 @@ namespace CSRedis
                 args.Add(dir.ToString().ToUpperInvariant());
             if (isAlpha.HasValue && isAlpha.Value)
                 args.Add("ALPHA");
-            return new RedisArray.Generic<Dictionary<string,string>>(new RedisHash("SORT", args.ToArray()));
+            return new RedisArray.Generic<Dictionary<string, string>>(new RedisHash("SORT", args.ToArray()));
         }
         public static RedisArray.Strings Sort(string key, long? offset = null, long? count = null, string by = null, RedisSortDir? dir = null, bool? isAlpha = null, params string[] get)
         {
@@ -197,7 +211,7 @@ namespace CSRedis
             return new RedisScanCommand<string>(
                 new RedisArray.Strings("SCAN", args.ToArray()));
         }
-		public static RedisScanCommand<byte[]> ScanBytes(long cursor, string pattern = null, long? count = null)
+        public static RedisScanCommand<byte[]> ScanBytes(long cursor, string pattern = null, long? count = null)
         {
             var args = new List<object>();
             args.Add(cursor);
@@ -211,9 +225,14 @@ namespace CSRedis
         #endregion
 
         #region Hashes
+        public static RedisInt HStrLen(string key, string field)
+        {
+            object[] args = RedisArgs.Concat(key, field);
+            return new RedisInt("HSTRLEN", args);
+        }
         public static RedisInt HDel(string key, params string[] fields)
         {
-			object[] args = RedisArgs.Concat(key, fields);
+            object[] args = RedisArgs.Concat(key, fields);
             return new RedisInt("HDEL", args);
         }
         public static RedisBool HExists(string key, string field)
@@ -224,7 +243,7 @@ namespace CSRedis
         {
             return new RedisString("HGET", key, field);
         }
-		public static RedisBytes HGetBytes(string key, string field)
+        public static RedisBytes HGetBytes(string key, string field)
         {
             return new RedisBytes("HGET", key, field);
         }
@@ -237,7 +256,7 @@ namespace CSRedis
         {
             return new RedisHash("HGETALL", key);
         }
-		public static RedisHashBytes HGetAllBytes(string key)
+        public static RedisHashBytes HGetAllBytes(string key)
         {
             return new RedisHashBytes("HGETALL", key);
         }
@@ -245,7 +264,7 @@ namespace CSRedis
         {
             return new RedisInt("HINCRBY", key, field, increment);
         }
-        public static RedisFloat HIncrByFloat(string key, string field, double increment)
+        public static RedisFloat HIncrByFloat(string key, string field, decimal increment)
         {
             return new RedisFloat("HINCRBYFLOAT", key, field, increment);
         }
@@ -259,12 +278,12 @@ namespace CSRedis
         }
         public static RedisArray.Strings HMGet(string key, params string[] fields)
         {
-			object[] args = RedisArgs.Concat(key, fields);
+            object[] args = RedisArgs.Concat(key, fields);
             return new RedisArray.Strings("HMGET", args);
         }
-		public static RedisArray.Bytes HMGetBytes(string key, params string[] fields)
+        public static RedisArray.Bytes HMGetBytes(string key, params string[] fields)
         {
-			object[] args = RedisArgs.Concat(key, fields);
+            object[] args = RedisArgs.Concat(key, fields);
             return new RedisArray.Bytes("HMGET", args);
         }
         public static RedisStatus HMSet(string key, Dictionary<string, object> dict)
@@ -302,7 +321,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("HVALS", key);
         }
-		public static RedisArray.Bytes HValsBytes(string key)
+        public static RedisArray.Bytes HValsBytes(string key)
         {
             return new RedisArray.Bytes("HVALS", key);
         }
@@ -317,10 +336,10 @@ namespace CSRedis
                 args.AddRange(new object[] { "COUNT", count });
             return new RedisScanCommand<Tuple<string, string>>(
                 new RedisArray.WeakPairs<string, string>("HSCAN", args.ToArray()));
-                //new RedisArray.Generic<Tuple<string, string>>(
-                    //new RedisTuple.Generic<string, string>.Bulk("HSCAN", args.ToArray())));
+            //new RedisArray.Generic<Tuple<string, string>>(
+            //new RedisTuple.Generic<string, string>.Bulk("HSCAN", args.ToArray())));
         }
-		public static RedisScanCommand<Tuple<string, byte[]>> HScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        public static RedisScanCommand<Tuple<string, byte[]>> HScanBytes(string key, long cursor, string pattern = null, long? count = null)
         {
             var args = new List<object>();
             args.Add(key);
@@ -331,80 +350,81 @@ namespace CSRedis
                 args.AddRange(new object[] { "COUNT", count });
             return new RedisScanCommand<Tuple<string, byte[]>>(
                 new RedisArray.StrongPairs<string, byte[]>(new RedisString(null), new RedisBytes(null), "HSCAN", args.ToArray()));
-                //new RedisArray.Generic<Tuple<string, string>>(
-                    //new RedisTuple.Generic<string, string>.Bulk("HSCAN", args.ToArray())));
+            //new RedisArray.Generic<Tuple<string, string>>(
+            //new RedisTuple.Generic<string, string>.Bulk("HSCAN", args.ToArray())));
         }
         #endregion
 
         #region Lists
         public static RedisTuple BLPopWithKey(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisTuple("BLPOP", args);
         }
-		public static RedisTuple.Generic<string, byte[]> BLPopBytesWithKey(int timeout, params string[] keys)
+        public static RedisTuple.Generic<string, byte[]> BLPopBytesWithKey(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisTuple.Generic<string, byte[]>.Single(new RedisString(null), new RedisBytes(null), "BLPOP", args);
         }
         public static RedisTuple BLPopWithKey(TimeSpan timeout, params string[] keys)
         {
             return BLPopWithKey((int)timeout.TotalSeconds, keys);
         }
-		public static RedisTuple.Generic<string, byte[]> BLPopBytesWithKey(TimeSpan timeout, params string[] keys)
+        public static RedisTuple.Generic<string, byte[]> BLPopBytesWithKey(TimeSpan timeout, params string[] keys)
         {
-			return BLPopBytesWithKey((int)timeout.TotalSeconds, keys);
-		}
+            return BLPopBytesWithKey((int)timeout.TotalSeconds, keys);
+        }
         public static RedisArray.IndexOf<string> BLPop(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisArray.IndexOf<string>(new RedisString("BLPOP", args), 1);
         }
-		public static RedisArray.IndexOf<byte[]> BLPopBytes(int timeout, params string[] keys)
+        public static RedisArray.IndexOf<byte[]> BLPopBytes(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisArray.IndexOf<byte[]>(new RedisBytes("BLPOP", args), 1);
         }
         public static RedisArray.IndexOf<string> BLPop(TimeSpan timeout, params string[] keys)
         {
             return BLPop((int)timeout.TotalSeconds, keys);
         }
-		public static RedisArray.IndexOf<byte[]> BLPopBytes(TimeSpan timeout, params string[] keys)
+        public static RedisArray.IndexOf<byte[]> BLPopBytes(TimeSpan timeout, params string[] keys)
         {
             return BLPopBytes((int)timeout.TotalSeconds, keys);
         }
         public static RedisTuple BRPopWithKey(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisTuple("BRPOP", args);
         }
-		public static RedisTuple.Generic<string, byte[]> BRPopBytesWithKey(int timeout, params string[] keys) {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
-			return new RedisTuple.Generic<string, byte[]>.Single(new RedisString(null), new RedisBytes(null), "BRPOP", args);
-		}
-		public static RedisTuple BRPopWithKey(TimeSpan timeout, params string[] keys)
+        public static RedisTuple.Generic<string, byte[]> BRPopBytesWithKey(int timeout, params string[] keys)
+        {
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            return new RedisTuple.Generic<string, byte[]>.Single(new RedisString(null), new RedisBytes(null), "BRPOP", args);
+        }
+        public static RedisTuple BRPopWithKey(TimeSpan timeout, params string[] keys)
         {
             return BRPopWithKey((int)timeout.TotalSeconds, keys);
         }
-		public static RedisTuple.Generic<string, byte[]> BRPopBytesWithKey(TimeSpan timeout, params string[] keys)
+        public static RedisTuple.Generic<string, byte[]> BRPopBytesWithKey(TimeSpan timeout, params string[] keys)
         {
             return BRPopBytesWithKey((int)timeout.TotalSeconds, keys);
         }
         public static RedisArray.IndexOf<string> BRPop(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisArray.IndexOf<string>(new RedisString("BRPOP", args), 1);
         }
-		public static RedisArray.IndexOf<byte[]> BRPopBytes(int timeout, params string[] keys)
+        public static RedisArray.IndexOf<byte[]> BRPopBytes(int timeout, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(keys, new object[] { timeout });
+            object[] args = RedisArgs.Concat(keys, new object[] { timeout });
             return new RedisArray.IndexOf<byte[]>(new RedisBytes("BRPOP", args), 1);
         }
         public static RedisArray.IndexOf<string> BRPop(TimeSpan timeout, params string[] keys)
         {
             return BRPop((int)timeout.TotalSeconds, keys);
         }
-		public static RedisArray.IndexOf<byte[]> BRPopBytes(TimeSpan timeout, params string[] keys)
+        public static RedisArray.IndexOf<byte[]> BRPopBytes(TimeSpan timeout, params string[] keys)
         {
             return BRPopBytes((int)timeout.TotalSeconds, keys);
         }
@@ -412,7 +432,7 @@ namespace CSRedis
         {
             return new RedisString.Nullable("BRPOPLPUSH", source, destination, timeout);
         }
-		public static RedisBytes BRPopBytesLPush(string source, string destination, int timeout)
+        public static RedisBytes BRPopBytesLPush(string source, string destination, int timeout)
         {
             return new RedisBytes("BRPOPLPUSH", source, destination, timeout);
         }
@@ -420,7 +440,7 @@ namespace CSRedis
         {
             return BRPopLPush(source, destination, (int)timeout.TotalSeconds);
         }
-		public static RedisBytes BRPopBytesLPush(string source, string destination, TimeSpan timeout)
+        public static RedisBytes BRPopBytesLPush(string source, string destination, TimeSpan timeout)
         {
             return BRPopBytesLPush(source, destination, (int)timeout.TotalSeconds);
         }
@@ -428,7 +448,7 @@ namespace CSRedis
         {
             return new RedisString("LINDEX", key, index);
         }
-		public static RedisBytes LIndexBytes(string key, long index)
+        public static RedisBytes LIndexBytes(string key, long index)
         {
             return new RedisBytes("LINDEX", key, index);
         }
@@ -444,13 +464,13 @@ namespace CSRedis
         {
             return new RedisString("LPOP", key);
         }
-		public static RedisBytes LPopBytes(string key)
+        public static RedisBytes LPopBytes(string key)
         {
             return new RedisBytes("LPOP", key);
         }
         public static RedisInt LPush(string key, params object[] values)
         {
-			object[] args = RedisArgs.Concat(new[] { key }, values);
+            object[] args = RedisArgs.Concat(new[] { key }, values);
             return new RedisInt("LPUSH", args);
         }
         public static RedisInt LPushX(string key, object value)
@@ -461,7 +481,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("LRANGE", key, start, stop);
         }
-		public static RedisArray.Bytes LRangeBytes(string key, long start, long stop)
+        public static RedisArray.Bytes LRangeBytes(string key, long start, long stop)
         {
             return new RedisArray.Bytes("LRANGE", key, start, stop);
         }
@@ -481,7 +501,7 @@ namespace CSRedis
         {
             return new RedisString("RPOP", key);
         }
-		public static RedisBytes RPopBytes(string key)
+        public static RedisBytes RPopBytes(string key)
         {
             return new RedisBytes("RPOP", key);
         }
@@ -489,18 +509,18 @@ namespace CSRedis
         {
             return new RedisString("RPOPLPUSH", source, destination);
         }
-		public static RedisBytes RPopBytesLPush(string source, string destination)
+        public static RedisBytes RPopBytesLPush(string source, string destination)
         {
             return new RedisBytes("RPOPLPUSH", source, destination);
         }
         public static RedisInt RPush(string key, params object[] values)
         {
-			object[] args = RedisArgs.Concat(key, values);
+            object[] args = RedisArgs.Concat(key, values);
             return new RedisInt("RPUSH", args);
         }
         public static RedisInt RPushX(string key, object value)
         {
-			return new RedisInt("RPUSHX", key, value);
+            return new RedisInt("RPUSHX", key, value);
         }
         #endregion
 
@@ -518,7 +538,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("SDIFF", keys);
         }
-		public static RedisArray.Bytes SDiffBytes(params string[] keys)
+        public static RedisArray.Bytes SDiffBytes(params string[] keys)
         {
             return new RedisArray.Bytes("SDIFF", keys);
         }
@@ -531,7 +551,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("SINTER", keys);
         }
-		public static RedisArray.Bytes SInterBytes(params string[] keys)
+        public static RedisArray.Bytes SInterBytes(params string[] keys)
         {
             return new RedisArray.Bytes("SINTER", keys);
         }
@@ -540,7 +560,7 @@ namespace CSRedis
             object[] args = RedisArgs.Concat(destination, keys);
             return new RedisInt("SINTERSTORE", args);
         }
-        public static RedisBool SIsMember(string key, object member) 
+        public static RedisBool SIsMember(string key, object member)
         {
             return new RedisBool("SISMEMBER", key, member);
         }
@@ -548,7 +568,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("SMEMBERS", key);
         }
-		public static RedisArray.Bytes SMembersBytes(string key)
+        public static RedisArray.Bytes SMembersBytes(string key)
         {
             return new RedisArray.Bytes("SMEMBERS", key);
         }
@@ -560,26 +580,35 @@ namespace CSRedis
         {
             return new RedisString("SPOP", key);
         }
-		public static RedisBytes SPopBytes(string key)
+        public static RedisBytes SPopBytes(string key)
         {
             return new RedisBytes("SPOP", key);
+        }
+        public static RedisArray.Strings SPop(string key, long count)
+        {
+            return new RedisArray.Strings("SPOP", key, count);
+        }
+        public static RedisArray.Bytes SPopBytes(string key, long count)
+        {
+            return new RedisArray.Bytes("SPOP", key, count);
         }
         public static RedisString SRandMember(string key)
         {
             return new RedisString("SRANDMEMBER", key);
         }
-		public static RedisBytes SRandMemberBytes(string key) {
-			return new RedisBytes("SRANDMEMBER", key);
-		}
-		public static RedisArray.Strings SRandMembers(string key, long count)
+        public static RedisBytes SRandMemberBytes(string key)
+        {
+            return new RedisBytes("SRANDMEMBER", key);
+        }
+        public static RedisArray.Strings SRandMembers(string key, long count)
         {
             return new RedisArray.Strings("SRANDMEMBER", key, count);
         }
-		public static RedisArray.Bytes SRandMembersBytes(string key, long count)
+        public static RedisArray.Bytes SRandMembersBytes(string key, long count)
         {
             return new RedisArray.Bytes("SRANDMEMBER", key, count);
         }
-        public static RedisInt SRem(string key, params object[] members) 
+        public static RedisInt SRem(string key, params object[] members)
         {
             object[] args = RedisArgs.Concat(key, members);
             return new RedisInt("SREM", args);
@@ -588,13 +617,13 @@ namespace CSRedis
         {
             return new RedisArray.Strings("SUNION", keys);
         }
-		public static RedisArray.Bytes SUnionBytes(params string[] keys)
+        public static RedisArray.Bytes SUnionBytes(params string[] keys)
         {
             return new RedisArray.Bytes("SUNION", keys);
         }
         public static RedisInt SUnionStore(string destination, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(destination, keys);
+            object[] args = RedisArgs.Concat(destination, keys);
             return new RedisInt("SUNIONSTORE", args);
         }
         public static RedisScanCommand<string> SScan(string key, long cursor, string pattern = null, long? count = null)
@@ -609,7 +638,7 @@ namespace CSRedis
             return new RedisScanCommand<string>(
                 new RedisArray.Strings("SSCAN", args.ToArray()));
         }
-		public static RedisScanCommand<byte[]> SScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        public static RedisScanCommand<byte[]> SScanBytes(string key, long cursor, string pattern = null, long? count = null)
         {
             var args = new List<object>();
             args.Add(key);
@@ -624,6 +653,23 @@ namespace CSRedis
         #endregion
 
         #region Sorted Sets
+        public static RedisArray.WeakPairs<string, decimal> ZPopMax(string key, long count)
+        {
+            return new RedisArray.WeakPairs<string, decimal>("ZPOPMAX", key, count);
+        }
+        public static RedisArray.StrongPairs<byte[], decimal> ZPopMaxBytes(string key, long count)
+        {
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZPOPMAX", key, count);
+        }
+        public static RedisArray.WeakPairs<string, decimal> ZPopMin(string key, long count)
+        {
+            return new RedisArray.WeakPairs<string, decimal>("ZPOPMIN", key, count);
+        }
+        public static RedisArray.StrongPairs<byte[], decimal> ZPopMinBytes(string key, long count)
+        {
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZPOPMIN", key, count);
+        }
+
         public static RedisInt ZAdd<TScore, TMember>(string key, params Tuple<TScore, TMember>[] scoreMembers)
         {
             object[] args = RedisArgs.Concat(key, RedisArgs.GetTupleArgs(scoreMembers));
@@ -638,7 +684,7 @@ namespace CSRedis
         {
             return new RedisInt("ZCARD", key);
         }
-        public static RedisInt ZCount(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)
+        public static RedisInt ZCount(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
@@ -648,11 +694,11 @@ namespace CSRedis
         {
             return new RedisInt("ZCOUNT", key, min, max);
         }
-        public static RedisFloat ZIncrBy(string key, double increment, string member)
+        public static RedisFloat ZIncrBy(string key, decimal increment, object member)
         {
             return new RedisFloat("ZINCRBY", key, increment, member);
         }
-        public static RedisInt ZInterStore(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public static RedisInt ZInterStore(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             List<object> args = new List<object>();
             args.Add(destination);
@@ -678,47 +724,48 @@ namespace CSRedis
                 : new[] { key, start.ToString(), stop.ToString() };
             return new RedisArray.Strings("ZRANGE", args);
         }
-		public static RedisArray.Bytes ZRangeBytes(string key, long start, long stop, bool withScores = false)
+        public static RedisArray.Bytes ZRangeBytes(string key, long start, long stop, bool withScores = false)
         {
             string[] args = withScores
                 ? new[] { key, start.ToString(), stop.ToString(), "WITHSCORES" }
                 : new[] { key, start.ToString(), stop.ToString() };
             return new RedisArray.Bytes("ZRANGE", args);
         }
-        public static RedisArray.WeakPairs<string, double> ZRangeWithScores(string key, long start, long stop)
+        public static RedisArray.WeakPairs<string, decimal> ZRangeWithScores(string key, long start, long stop)
         {
-            return new RedisArray.WeakPairs<string, double>("ZRANGE", key, start, stop, "WITHSCORES");
+            return new RedisArray.WeakPairs<string, decimal>("ZRANGE", key, start, stop, "WITHSCORES");
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRangeBytesWithScores(string key, long start, long stop)
+        public static RedisArray.StrongPairs<byte[], decimal> ZRangeBytesWithScores(string key, long start, long stop)
         {
-            return new RedisArray.StrongPairs<byte[], double>(new RedisBytes(null), new RedisFloat(null), "ZRANGE", key, start, stop, "WITHSCORES");
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZRANGE", key, start, stop, "WITHSCORES");
         }
-        public static RedisArray.Strings ZRangeByScore(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public static RedisArray.Strings ZRangeByScore(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRangeByScore(key, min_score, max_score, withScores, offset, count);
         }
-		public static RedisArray.Bytes ZRangeBytesByScore(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public static RedisArray.Bytes ZRangeBytesByScore(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRangeBytesByScore(key, min_score, max_score, withScores, offset, count);
         }
-        public static RedisArray.WeakPairs<string, double> ZRangeByScoreWithScores(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public static RedisArray.WeakPairs<string, decimal> ZRangeByScoreWithScores(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRangeByScoreWithScores(key, min_score, max_score, offset, count);
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRangeBytesByScoreWithScores(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null) {
-			string min_score = RedisArgs.GetScore(min, exclusiveMin);
-			string max_score = RedisArgs.GetScore(max, exclusiveMax);
-			return ZRangeBytesByScoreWithScores(key, min_score, max_score, offset, count);
-		}
-		public static RedisArray.Strings ZRangeByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
+        public static RedisArray.StrongPairs<byte[], decimal> ZRangeBytesByScoreWithScores(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, min, max};
+            string min_score = RedisArgs.GetScore(min, exclusiveMin);
+            string max_score = RedisArgs.GetScore(max, exclusiveMax);
+            return ZRangeBytesByScoreWithScores(key, min_score, max_score, offset, count);
+        }
+        public static RedisArray.Strings ZRangeByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
+        {
+            object[] args = new[] { key, min, max };
             if (withScores)
                 args = RedisArgs.Concat(args, new[] { "WITHSCORES" });
             if (offset.HasValue && count.HasValue)
@@ -726,9 +773,9 @@ namespace CSRedis
 
             return new RedisArray.Strings("ZRANGEBYSCORE", args);
         }
-		public static RedisArray.Bytes ZRangeBytesByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
+        public static RedisArray.Bytes ZRangeBytesByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, min, max};
+            object[] args = new[] { key, min, max };
             if (withScores)
                 args = RedisArgs.Concat(args, new[] { "WITHSCORES" });
             if (offset.HasValue && count.HasValue)
@@ -736,42 +783,43 @@ namespace CSRedis
 
             return new RedisArray.Bytes("ZRANGEBYSCORE", args);
         }
-        public static RedisArray.WeakPairs<string, double> ZRangeByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null) 
+        public static RedisArray.WeakPairs<string, decimal> ZRangeByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, min, max, "WITHSCORES" };
+            object[] args = new[] { key, min, max, "WITHSCORES" };
             if (offset.HasValue && count.HasValue)
                 args = RedisArgs.Concat(args, new[] { "LIMIT", offset.Value.ToString(), count.Value.ToString() });
 
-            return new RedisArray.WeakPairs<string, double>("ZRANGEBYSCORE", args);
+            return new RedisArray.WeakPairs<string, decimal>("ZRANGEBYSCORE", args);
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRangeBytesByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null) {
-			object[] args = new[] { key, min, max, "WITHSCORES" };
-			if (offset.HasValue && count.HasValue)
-				args = RedisArgs.Concat(args, new[] { "LIMIT", offset.Value.ToString(), count.Value.ToString() });
+        public static RedisArray.StrongPairs<byte[], decimal> ZRangeBytesByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
+        {
+            object[] args = new[] { key, min, max, "WITHSCORES" };
+            if (offset.HasValue && count.HasValue)
+                args = RedisArgs.Concat(args, new[] { "LIMIT", offset.Value.ToString(), count.Value.ToString() });
 
-			return new RedisArray.StrongPairs<byte[], double>(new RedisBytes(null), new RedisFloat(null), "ZRANGEBYSCORE", args);
-		}
-		public static RedisInt.Nullable ZRank(string key, object member) 
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZRANGEBYSCORE", args);
+        }
+        public static RedisInt.Nullable ZRank(string key, object member)
         {
             return new RedisInt.Nullable("ZRANK", key, member);
         }
-        public static RedisInt ZRem(string key, params object[] members) 
+        public static RedisInt ZRem(string key, params object[] members)
         {
-			object[] args = RedisArgs.Concat(new[] { key }, members);
+            object[] args = RedisArgs.Concat(new[] { key }, members);
             return new RedisInt("ZREM", args);
         }
         public static RedisInt ZRemRangeByRank(string key, long start, long stop)
         {
             return new RedisInt("ZREMRANGEBYRANK", key, start, stop);
         }
-        public static RedisInt ZRemRangeByScore(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)  
+        public static RedisInt ZRemRangeByScore(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
 
             return ZRemRangeByScore(key, min_score, max_score);
         }
-		public static RedisInt ZRemRangeByScore(string key, string min, string max)  
+        public static RedisInt ZRemRangeByScore(string key, string min, string max)
         {
             return new RedisInt("ZREMRANGEBYSCORE", key, min, max);
         }
@@ -782,28 +830,28 @@ namespace CSRedis
                 : new[] { key, start.ToString(), stop.ToString() };
             return new RedisArray.Strings("ZREVRANGE", args);
         }
-		public static RedisArray.Bytes ZRevRangeBytes(string key, long start, long stop, bool withScores = false)
+        public static RedisArray.Bytes ZRevRangeBytes(string key, long start, long stop, bool withScores = false)
         {
             string[] args = withScores
                 ? new[] { key, start.ToString(), stop.ToString(), "WITHSCORES" }
                 : new[] { key, start.ToString(), stop.ToString() };
             return new RedisArray.Bytes("ZREVRANGE", args);
         }
-        public static RedisArray.WeakPairs<string, double> ZRevRangeWithScores(string key, long start, long stop)
+        public static RedisArray.WeakPairs<string, decimal> ZRevRangeWithScores(string key, long start, long stop)
         {
-            return new RedisArray.WeakPairs<string, double>("ZREVRANGE", key, start.ToString(), stop.ToString(), "WITHSCORES");
+            return new RedisArray.WeakPairs<string, decimal>("ZREVRANGE", key, start.ToString(), stop.ToString(), "WITHSCORES");
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRevRangeBytesWithScores(string key, long start, long stop)
+        public static RedisArray.StrongPairs<byte[], decimal> ZRevRangeBytesWithScores(string key, long start, long stop)
         {
-            return new RedisArray.StrongPairs<byte[], double>(new RedisBytes(null), new RedisFloat(null), "ZREVRANGE", key, start.ToString(), stop.ToString(), "WITHSCORES");
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZREVRANGE", key, start.ToString(), stop.ToString(), "WITHSCORES");
         }
-        public static RedisArray.Strings ZRevRangeByScore(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public static RedisArray.Strings ZRevRangeByScore(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRevRangeByScore(key, max_score, min_score, withScores, offset, count);
         }
-		public static RedisArray.Bytes ZRevRangeBytesByScore(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public static RedisArray.Bytes ZRevRangeBytesByScore(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
@@ -811,7 +859,7 @@ namespace CSRedis
         }
         public static RedisArray.Strings ZRevRangeByScore(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, max, min };
+            object[] args = new[] { key, max, min };
             if (withScores)
                 args = RedisArgs.Concat(args, new[] { "WITHSCORES" });
             if (count.HasValue)
@@ -819,9 +867,9 @@ namespace CSRedis
 
             return new RedisArray.Strings("ZREVRANGEBYSCORE", args);
         }
-		public static RedisArray.Bytes ZRevRangeBytesByScore(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
+        public static RedisArray.Bytes ZRevRangeBytesByScore(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, max, min };
+            object[] args = new[] { key, max, min };
             if (withScores)
                 args = RedisArgs.Concat(args, new[] { "WITHSCORES" });
             if (count.HasValue)
@@ -829,33 +877,33 @@ namespace CSRedis
 
             return new RedisArray.Bytes("ZREVRANGEBYSCORE", args);
         }
-        public static RedisArray.WeakPairs<string, double> ZRevRangeByScoreWithScores(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public static RedisArray.WeakPairs<string, decimal> ZRevRangeByScoreWithScores(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRevRangeByScoreWithScores(key, max_score, min_score, offset, count);
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRevRangeBytesByScoreWithScores(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public static RedisArray.StrongPairs<byte[], decimal> ZRevRangeBytesByScoreWithScores(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             string min_score = RedisArgs.GetScore(min, exclusiveMin);
             string max_score = RedisArgs.GetScore(max, exclusiveMax);
             return ZRevRangeBytesByScoreWithScores(key, max_score, min_score, offset, count);
         }
-        public static RedisArray.WeakPairs<string, double> ZRevRangeByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
+        public static RedisArray.WeakPairs<string, decimal> ZRevRangeByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, max, min, "WITHSCORES" };
+            object[] args = new[] { key, max, min, "WITHSCORES" };
             if (count.HasValue)
                 args = RedisArgs.Concat(args, new[] { "LIMIT", (offset ?? 0).ToString(), count.Value.ToString() });
 
-            return new RedisArray.WeakPairs<string, double>("ZREVRANGEBYSCORE", args);
+            return new RedisArray.WeakPairs<string, decimal>("ZREVRANGEBYSCORE", args);
         }
-		public static RedisArray.StrongPairs<byte[], double> ZRevRangeBytesByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
+        public static RedisArray.StrongPairs<byte[], decimal> ZRevRangeBytesByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
         {
-			object[] args = new[] { key, max, min, "WITHSCORES" };
+            object[] args = new[] { key, max, min, "WITHSCORES" };
             if (count.HasValue)
                 args = RedisArgs.Concat(args, new[] { "LIMIT", (offset ?? 0).ToString(), count.Value.ToString() });
 
-            return new RedisArray.StrongPairs<byte[], double>(new RedisBytes(null), new RedisFloat(null), "ZREVRANGEBYSCORE", args);
+            return new RedisArray.StrongPairs<byte[], decimal>(new RedisBytes(null), new RedisFloat(null), "ZREVRANGEBYSCORE", args);
         }
         public static RedisInt.Nullable ZRevRank(string key, object member)
         {
@@ -865,7 +913,7 @@ namespace CSRedis
         {
             return new RedisFloat.Nullable("ZSCORE", key, member);
         }
-        public static RedisInt ZUnionStore(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public static RedisInt ZUnionStore(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             List<object> args = new List<object>();
             args.Add(destination);
@@ -884,7 +932,7 @@ namespace CSRedis
             }
             return new RedisInt("ZUNIONSTORE", args.ToArray());
         }
-        public static RedisScanCommand<Tuple<string, double>> ZScan(string key, long cursor, string pattern = null, long? count = null)
+        public static RedisScanCommand<Tuple<string, decimal>> ZScan(string key, long cursor, string pattern = null, long? count = null)
         {
             var args = new List<object>();
             args.Add(key);
@@ -893,26 +941,27 @@ namespace CSRedis
                 args.AddRange(new[] { "MATCH", pattern });
             if (count != null)
                 args.AddRange(new object[] { "COUNT", count });
-            return new RedisScanCommand<Tuple<string, double>>(
-                new RedisArray.WeakPairs<string, double>("ZSCAN", args.ToArray()));
-                    //<Tuple<string, double>>(
-                    //new RedisTuple.Generic<string, double>.Bulk("ZSCAN", args.ToArray())));
+            return new RedisScanCommand<Tuple<string, decimal>>(
+                new RedisArray.WeakPairs<string, decimal>("ZSCAN", args.ToArray()));
+            //<Tuple<string, decimal>>(
+            //new RedisTuple.Generic<string, decimal>.Bulk("ZSCAN", args.ToArray())));
         }
-		public static RedisScanCommand<Tuple<byte[], double>> ZScanBytes(string key, long cursor, string pattern = null, long? count = null) {
-			var args = new List<object>();
-			args.Add(key);
-			args.Add(cursor);
-			if (pattern != null)
-				args.AddRange(new[] { "MATCH", pattern });
-			if (count != null)
-				args.AddRange(new object[] { "COUNT", count });
-			return new RedisScanCommand<Tuple<byte[], double>>(
-				new RedisArray.StrongPairs<byte[], double>(
-				new RedisBytes(null), new RedisFloat(null), "ZSCAN", args.ToArray()));
-			//<Tuple<string, double>>(
-			//new RedisTuple.Generic<string, double>.Bulk("ZSCAN", args.ToArray())));
-		}
-		public static RedisArray.Strings ZRangeByLex(string key, string min, string max, long? offset = null, long? count = null)
+        public static RedisScanCommand<Tuple<byte[], decimal>> ZScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        {
+            var args = new List<object>();
+            args.Add(key);
+            args.Add(cursor);
+            if (pattern != null)
+                args.AddRange(new[] { "MATCH", pattern });
+            if (count != null)
+                args.AddRange(new object[] { "COUNT", count });
+            return new RedisScanCommand<Tuple<byte[], decimal>>(
+                new RedisArray.StrongPairs<byte[], decimal>(
+                new RedisBytes(null), new RedisFloat(null), "ZSCAN", args.ToArray()));
+            //<Tuple<string, decimal>>(
+            //new RedisTuple.Generic<string, decimal>.Bulk("ZSCAN", args.ToArray())));
+        }
+        public static RedisArray.Strings ZRangeByLex(string key, string min, string max, long? offset = null, long? count = null)
         {
             List<object> args = new List<object>();
             args.Add(key);
@@ -922,7 +971,7 @@ namespace CSRedis
                 args.AddRange(new object[] { "LIMIT", offset, count });
             return new RedisArray.Strings("ZRANGEBYLEX", args.ToArray());
         }
-		public static RedisArray.Bytes ZRangeBytesByLex(string key, string min, string max, long? offset = null, long? count = null)
+        public static RedisArray.Bytes ZRangeBytesByLex(string key, string min, string max, long? offset = null, long? count = null)
         {
             List<object> args = new List<object>();
             args.Add(key);
@@ -961,7 +1010,7 @@ namespace CSRedis
         }
         public static RedisArray.StrongPairs<string, long> PubSubNumSub(params string[] channels)
         {
-			object[] args = RedisArgs.Concat("NUMSUB", channels);
+            object[] args = RedisArgs.Concat("NUMSUB", channels);
             return new RedisArray.StrongPairs<string, long>(
                 new RedisString(null), new RedisInt(null), "PUBSUB", args);
         }
@@ -986,12 +1035,12 @@ namespace CSRedis
         #region Scripting
         public static RedisObject.Strings Eval(string script, string[] keys, params object[] arguments)
         {
-			object[] args = RedisArgs.Concat(new object[] { script, keys.Length }, keys, arguments);
+            object[] args = RedisArgs.Concat(new object[] { script, keys.Length }, keys, arguments);
             return new RedisObject.Strings("EVAL", args);
         }
         public static RedisObject.Strings EvalSHA(string sha1, string[] keys, params object[] arguments)
         {
-			object[] args = RedisArgs.Concat(new object[] { sha1, keys.Length }, keys, arguments);
+            object[] args = RedisArgs.Concat(new object[] { sha1, keys.Length }, keys, arguments);
             return new RedisObject.Strings("EVALSHA", args);
         }
         public static RedisArray.Generic<bool> ScriptExists(params string[] sha1s)
@@ -1026,7 +1075,7 @@ namespace CSRedis
         }
         public static RedisInt BitOp(RedisBitOp operation, string destKey, params string[] keys)
         {
-			object[] args = RedisArgs.Concat(new[] { operation.ToString().ToUpperInvariant(), destKey }, keys);
+            object[] args = RedisArgs.Concat(new[] { operation.ToString().ToUpperInvariant(), destKey }, keys);
             return new RedisInt("BITOP", args);
         }
         public static RedisInt BitPos(string key, bool bit, long? start = null, long? end = null)
@@ -1057,18 +1106,19 @@ namespace CSRedis
         {
             return new RedisString("GET", key);
         }
-		public static RedisBytes GetBytes (string key) {
-			return new RedisBytes("GET", key);
-		}
-		public static RedisBool GetBit(string key, uint offset)
+        public static RedisBytes GetBytes(string key)
+        {
+            return new RedisBytes("GET", key);
+        }
+        public static RedisBool GetBit(string key, uint offset)
         {
             return new RedisBool("GETBIT", key, offset);
         }
-        public static RedisString GetRange(string key, long start, long end) 
+        public static RedisString GetRange(string key, long start, long end)
         {
             return new RedisString("GETRANGE", key, start, end);
         }
-		public static RedisBytes GetRangeBytes(string key, long start, long end) 
+        public static RedisBytes GetRangeBytes(string key, long start, long end)
         {
             return new RedisBytes("GETRANGE", key, start, end);
         }
@@ -1076,11 +1126,11 @@ namespace CSRedis
         {
             return new RedisString("GETSET", key, value);
         }
-		public static RedisBytes GetSetBytes(string key, object value)
-		{
-			return new RedisBytes("GETSET", key, value);
-		}
-		public static RedisInt Incr(string key)
+        public static RedisBytes GetSetBytes(string key, object value)
+        {
+            return new RedisBytes("GETSET", key, value);
+        }
+        public static RedisInt Incr(string key)
         {
             return new RedisInt("INCR", key);
         }
@@ -1088,7 +1138,7 @@ namespace CSRedis
         {
             return new RedisInt("INCRBY", key, increment);
         }
-        public static RedisFloat IncrByFloat(string key, double increment)
+        public static RedisFloat IncrByFloat(string key, decimal increment)
         {
             return new RedisFloat("INCRBYFLOAT", key, increment);
         }
@@ -1096,7 +1146,7 @@ namespace CSRedis
         {
             return new RedisArray.Strings("MGET", keys);
         }
-		public static RedisArray.Bytes MGetBytes(params string[] keys)
+        public static RedisArray.Bytes MGetBytes(params string[] keys)
         {
             return new RedisArray.Bytes("MGET", keys);
         }
@@ -1126,7 +1176,7 @@ namespace CSRedis
         {
             return new RedisStatus("SET", key, value);
         }
-        public static RedisStatus.Nullable Set(string key, object value, TimeSpan expiration, RedisExistence? condition = null) 
+        public static RedisStatus.Nullable Set(string key, object value, TimeSpan expiration, RedisExistence? condition = null)
         {
             return Set(key, value, (long)expiration.TotalMilliseconds, condition);
         }
@@ -1341,7 +1391,7 @@ namespace CSRedis
         #region HyperLogLog
         public static RedisBool PfAdd(string key, params object[] elements)
         {
-			object[] args = RedisArgs.Concat(key, elements);
+            object[] args = RedisArgs.Concat(key, elements);
             return new RedisBool("PFADD", args);
         }
         public static RedisInt PfCount(params string[] keys)
@@ -1350,8 +1400,551 @@ namespace CSRedis
         }
         public static RedisStatus PfMerge(string destKey, params string[] sourceKeys)
         {
-			object[] args = RedisArgs.Concat(destKey, sourceKeys);
+            object[] args = RedisArgs.Concat(destKey, sourceKeys);
             return new RedisStatus("PFMERGE", args);
+        }
+        #endregion
+
+        #region Streams
+        public static RedisInt XAck(string key, string group, params string[] id)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, group });
+            args.AddRange(id.Select(a => (object)a));
+            return new RedisInt("XACK", args.ToArray());
+        }
+        public static RedisString XAdd(string key, long maxLen, string id = "*", params (string, string)[] fieldValues)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, id });
+            if (maxLen > 0) args.AddRange(new object[] { "MAXLEN", maxLen });
+            else if (maxLen < 0) args.AddRange(new object[] { "MAXLEN", $"~{Math.Abs(maxLen)}" });
+            args.AddRange(fieldValues.Select(a => new[] { a.Item1, a.Item2 }).SelectMany(a => a).Select(a => (object)a));
+            return new RedisString("XADD", args.ToArray());
+        }
+
+        public static RedisXRangeCommand XClaim(string key, string group, string consumer, long minIdleTime, params string[] id)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, group, consumer, minIdleTime });
+            args.AddRange(id.Select(a => (object)a));
+            return new RedisXRangeCommand("XCLAIM", args.ToArray());
+        }
+        public static RedisXRangeCommand XClaim(string key, string group, string consumer, long minIdleTime, string[] id, long idle, long retryCount, bool force)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, group, consumer, minIdleTime });
+            args.AddRange(id.Select(a => (object)a));
+            args.AddRange(new object[] { "IDLE", idle, "RETRYCOUNT", retryCount });
+            if (force) args.Add("FORCE");
+            return new RedisXRangeCommand("XCLAIM", args.ToArray());
+        }
+        public static RedisArray.Strings XClaimJustId(string key, string group, string consumer, long minIdleTime, params string[] id)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, group, consumer, minIdleTime });
+            args.AddRange(id.Select(a => (object)a));
+            args.Add("JUSTID");
+            return new RedisArray.Strings("XCLAIM", args.ToArray());
+        }
+        public static RedisArray.Strings XClaimJustId(string key, string group, string consumer, long minIdleTime, string[] id, long idle, long retryCount, bool force)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, group, consumer, minIdleTime });
+            args.AddRange(id.Select(a => (object)a));
+            args.AddRange(new object[] { "IDLE", idle, "RETRYCOUNT", retryCount });
+            if (force) args.Add("FORCE");
+            args.Add("JUSTID");
+            return new RedisArray.Strings("XCLAIM", args.ToArray());
+        }
+
+        public static RedisInt XDel(string key, params string[] id)
+        {
+            object[] args = RedisArgs.Concat(key, id);
+            return new RedisInt("XDEL", args);
+        }
+
+        public static class XGroup
+        {
+            public static RedisStatus Create(string key, string group, string id = "$", bool MkStream = false)
+            {
+                var args = new List<object>();
+                args.AddRange(new object[] { "CREATE", key, group, id, });
+                if (MkStream) args.Add("MKSTREAM");
+                return new RedisStatus("XGROUP", args.ToArray());
+            }
+            public static RedisStatus SetId(string key, string group, string id = "$")
+            {
+                return new RedisStatus("XGROUP", "SETID", key, group, id);
+            }
+            public static RedisBool Destroy(string key, string group)
+            {
+                return new RedisBool("XGROUP", "DESTROY", key, group);
+            }
+            public static RedisBool DelConsumer(string key, string group, string consumer)
+            {
+                return new RedisBool("XGROUP", "DELCONSUMER", key, group, consumer);
+            }
+        }
+
+        //XINFO
+        public static RedisXInfoStreamCommand XInfoStream(string key)
+        {
+            return new RedisXInfoStreamCommand("XINFO", "STREAM", key);
+        }
+        public static RedisXInfoGroupsCommand XInfoGroups(string key)
+        {
+            return new RedisXInfoGroupsCommand("XINFO", "GROUPS", key);
+        }
+        public static RedisXInfoConsumersCommand XInfoConsumers(string key, string group)
+        {
+            return new RedisXInfoConsumersCommand("XINFO", "CONSUMERS", key, group);
+        }
+
+        public static RedisInt XLen(string key)
+        {
+            return new RedisInt("XLEN", key);
+        }
+
+        //XPENDING
+        public static RedisXPendingCommand XPending(string key, string group)
+        {
+            return new RedisXPendingCommand("XPENDING", key, group);
+        }
+        public static RedisXPendingStartEndCountCommand XPending(string key, string group, string start, string end, long count, string consumer = null)
+        {
+            if (string.IsNullOrEmpty(consumer))
+                return new RedisXPendingStartEndCountCommand("XPENDING", key, group, start, end, count);
+            return new RedisXPendingStartEndCountCommand("XPENDING", key, group, start, end, count, consumer);
+        }
+
+        public static RedisXRangeCommand XRange(string key, string start, string end, long count = 1)
+        {
+            var args = new List<object>();
+            args.AddRange(new[] { key, start, end });
+            if (count > 0) args.AddRange(new[] { "COUNT", count.ToString() });
+
+            return new RedisXRangeCommand("XRANGE", args.ToArray());
+        }
+
+        public static RedisXRangeCommand XRevRange(string key, string end, string start, long count = 1)
+        {
+            var args = new List<object>();
+            args.AddRange(new[] { key, start, end });
+            if (count > 0) args.AddRange(new[] { "COUNT", count.ToString() });
+            return new RedisXRangeCommand("XREVRANGE", args.ToArray());
+        }
+
+        public static RedisXReadCommand XRead(long count, long block, params (string key, string id)[] streams)
+        {
+            var args = new List<object>();
+            if (count > 0) args.AddRange(new[] { "COUNT", count.ToString() });
+            if (block > 0) args.AddRange(new[] { "BLOCK", block.ToString() });
+            args.Add("STREAMS");
+            args.AddRange(streams.Select(a => (object)a.key));
+            args.AddRange(streams.Select(a => (object)a.id));
+            return new RedisXReadCommand("XREAD", args.ToArray());
+        }
+
+        public static RedisXReadCommand XReadGroup(string group, string consumer, long count, long block, params (string key, string id)[] streams)
+        {
+            var args = new List<object>();
+            args.AddRange(new[] { "GROUP", group, consumer });
+            if (count > 0) args.AddRange(new[] { "COUNT", count.ToString() });
+            if (block > 0) args.AddRange(new[] { "BLOCK", block.ToString() });
+            args.Add("STREAMS");
+            args.AddRange(streams.Select(a => (object)a.key));
+            args.AddRange(streams.Select(a => (object)a.id));
+            return new RedisXReadCommand("XREADGROUP", args.ToArray());
+        }
+
+        public static RedisInt XTrim(string key, long maxLen)
+        {
+            var maxLenArg = maxLen > 0 ? maxLen.ToString() : $"~{Math.Abs(maxLen)}";
+            return new RedisInt("XTRIM", key, "MAXLEN", maxLenArg);
+        }
+
+        #endregion
+
+        #region Bloom Filter
+        public static RedisStatus BfReserve(string key, decimal errorRate, long capacity, int expansion = 2, bool nonScaling = false)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, errorRate, capacity });
+            if (expansion != 2) args.AddRange(new object[] { "EXPANSION", expansion });
+            if (nonScaling) args.Add("NONSCALING");
+            return new RedisStatus("BF.RESERVE", args.ToArray());
+        }
+        public static RedisBool BfAdd(string key, object item)
+        {
+            return new RedisBool("BF.ADD", key, item);
+        }
+        public static RedisArray.Generic<bool> BfMAdd(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(items);
+            return new RedisArray.Generic<bool>(new RedisBool("BF.MADD", args.ToArray()));
+        }
+        public static RedisArray.Generic<bool> BfInsert(string key, object[] items, long? capacity = null, string error = null, int expansion = 2, bool noCreate = false, bool nonScaling = false)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            if (capacity != null) args.AddRange(new object[] { "CAPACITY", capacity });
+            if (string.IsNullOrEmpty(error) == false) args.AddRange(new object[] { "ERROR", error });
+            if (expansion != 2) args.Add(new object[] { "EXPANSION", expansion });
+            if (noCreate) args.Add("NOCREATE");
+            if (nonScaling) args.Add("NONSCALING");
+            args.Add("ITEMS");
+            args.AddRange(items);
+            return new RedisArray.Generic<bool>(new RedisBool("BF.INSERT", args.ToArray()));
+        }
+        public static RedisBool BfExists(string key, object item)
+        {
+            return new RedisBool("BF.EXISTS", key, item);
+        }
+        public static RedisArray.Generic<bool> BfMExists(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(items);
+            return new RedisArray.Generic<bool>(new RedisBool("BF.MEXISTS", args.ToArray()));
+        }
+        public static RedisScanCommand<byte[]> BfScanDump(string key, long iter)
+        {
+            return new RedisScanCommand<byte[]>(
+                   new RedisArray.Bytes("BF.SCANDUMP", key, iter));
+        }
+        public static RedisStatus BfLoadChunk(string key, long iter, byte[] data)
+        {
+            return new RedisStatus("BF.LOADCHUNK", key, iter, data);
+        }
+        public static RedisBfInfoCommand BfInfo(string key)
+        {
+            return new RedisBfInfoCommand("BF.INFO", key);
+        }
+        public class RedisBfInfoCommand : RedisCommand<(long capacity, long size, long numberOfFilters, long numberOfItemsInserted, long expansionRate)>
+        {
+            public RedisBfInfoCommand(string command, params object[] args)
+                : base(command, args)
+            {
+            }
+
+            public override (long capacity, long size, long numberOfFilters, long numberOfItemsInserted, long expansionRate) Parse(RedisReader reader)
+            {
+                long capacity = 0;
+                long size = 0;
+                long numberOfFilters = 0;
+                long numberOfItemsInserted = 0;
+                long expansionRate = 0;
+
+                reader.ExpectType(RedisMessage.MultiBulk);
+                long count = reader.ReadInt(false);
+                if (count % 2 != 0) throw new RedisProtocolException("BF.INFO 返回数据格式 1级 MultiBulk 长度应该为偶数");
+
+                for (var b = 0; b < count; b += 2)
+                {
+                    var key = reader.ReadBulkString().ToLower();
+                    switch (key)
+                    {
+                        case "capacity":
+                            capacity = reader.ReadInt();
+                            break;
+                        case "size":
+                            size = reader.ReadInt();
+                            break;
+                        case "number of filters":
+                            numberOfFilters = reader.ReadInt();
+                            break;
+                        case "number of items inserted":
+                            numberOfItemsInserted = reader.ReadInt();
+                            break;
+                        case "expansion rate":
+                            expansionRate = reader.ReadInt();
+                            break;
+                        default:
+                            reader.ReadBulkString();
+                            break;
+                    }
+                }
+                return (capacity, size, numberOfFilters, numberOfItemsInserted, expansionRate);
+            }
+        }
+        #endregion
+
+        #region RedisBloom Cuckoo Filter
+        public static RedisStatus CfReserve(string key, long capacity, long? bucketSize = null, long? maxIterations = null, int? expansion = null)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key, capacity });
+            if (bucketSize != 2) args.AddRange(new object[] { "BUCKETSIZE", bucketSize });
+            if (maxIterations != 2) args.AddRange(new object[] { "MAXITERATIONS", maxIterations });
+            if (expansion != 2) args.AddRange(new object[] { "EXPANSION", expansion });
+            return new RedisStatus("CF.RESERVE", args.ToArray());
+        }
+        public static RedisBool CfAdd(bool nx, string key, object item)
+        {
+            return new RedisBool(nx ? "CF.ADDNX": "CF.ADD", key, item);
+        }
+        public static RedisArray.Generic<bool> CfInsert(bool nx, string key, object[] items, long? capacity = null, bool noCreate = false)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            if (capacity != null) args.AddRange(new object[] { "CAPACITY", capacity });
+            if (noCreate) args.Add("NOCREATE");
+            args.Add("ITEMS");
+            args.AddRange(items);
+            return new RedisArray.Generic<bool>(new RedisBool(nx ? "CF.INSERTNX" : "CF.INSERT", args.ToArray()));
+        }
+        public static RedisBool CfExists(string key, object item)
+        {
+            return new RedisBool("CF.EXISTS", key, item);
+        }
+        public static RedisBool CfDel(string key, object item)
+        {
+            return new RedisBool("CF.DEL", key, item);
+        }
+        public static RedisInt CfCount(string key, object item)
+        {
+            return new RedisInt("CF.COUNT", key, item);
+        }
+        public static RedisScanCommand<byte[]> CfScanDump(string key, long iter)
+        {
+            return new RedisScanCommand<byte[]>(
+                   new RedisArray.Bytes("CF.SCANDUMP", key, iter));
+        }
+        public static RedisStatus CfLoadChunk(string key, long iter, byte[] data)
+        {
+            return new RedisStatus("CF.LOADCHUNK", key, iter, data);
+        }
+        public static RedisCfInfoCommand CfInfo(string key)
+        {
+            return new RedisCfInfoCommand("CF.INFO", key);
+        }
+        public class RedisCfInfoCommand : RedisCommand<(long size, long numberOfBuckets, long numberOfFilter, long numberOfItemsInserted, long numberOfItemsDeleted, long bucketSize, long expansionRate, long maxIteration)>
+        {
+            public RedisCfInfoCommand(string command, params object[] args)
+                : base(command, args)
+            {
+            }
+
+            public override (long size, long numberOfBuckets, long numberOfFilter, long numberOfItemsInserted, long numberOfItemsDeleted, long bucketSize, long expansionRate, long maxIteration) Parse(RedisReader reader)
+            {
+                long size = 0;
+                long numberOfBuckets = 0;
+                long numberOfFilter = 0;
+                long numberOfItemsInserted = 0;
+                long numberOfItemsDeleted = 0;
+                long bucketSize = 0;
+                long expansionRate = 0;
+                long maxIteration = 0;
+
+                reader.ExpectType(RedisMessage.MultiBulk);
+                long count = reader.ReadInt(false);
+                if (count % 2 != 0) throw new RedisProtocolException("CF.INFO 返回数据格式 1级 MultiBulk 长度应该为偶数");
+
+                for (var b = 0; b < count; b += 2)
+                {
+                    var key = reader.ReadBulkString().ToLower();
+                    switch (key)
+                    {
+                        case "size":
+                            size = reader.ReadInt();
+                            break;
+                        case "number of buckets":
+                            numberOfBuckets = reader.ReadInt();
+                            break;
+                        case "number of filter":
+                            numberOfFilter = reader.ReadInt();
+                            break;
+                        case "number of items inserted":
+                            numberOfItemsInserted = reader.ReadInt();
+                            break;
+                        case "number of items deleted":
+                            numberOfItemsDeleted = reader.ReadInt();
+                            break;
+                        case "bucket size":
+                            bucketSize = reader.ReadInt();
+                            break;
+                        case "expansion rate":
+                            expansionRate = reader.ReadInt();
+                            break;
+                        case "max iteration":
+                            maxIteration = reader.ReadInt();
+                            break;
+                        default:
+                            reader.ReadBulkString();
+                            break;
+                    }
+                }
+                return (size, numberOfBuckets, numberOfFilter, numberOfItemsInserted, numberOfItemsDeleted, bucketSize, expansionRate, maxIteration);
+            }
+        }
+        #endregion
+
+        #region RedisBloom Count-Min Sketch
+        public static RedisStatus CmsInitByDim(string key, long width, long depth)
+        {
+            return new RedisStatus("CMS.INITBYDIM", key, width, depth);
+        }
+        public static RedisStatus CmsInitByProb(string key, decimal error, decimal probability)
+        {
+            return new RedisStatus("CMS.INITBYPROB", key, error, probability);
+        }
+        public static RedisArray.Generic<long> CmsIncrBy(string key, params (object item, long increment)[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            foreach (var item in items) args.AddRange(new object[] { item.item, item.increment });
+            return new RedisArray.Generic<long>(new RedisInt("CMS.INCRBY", args.ToArray()));
+        }
+        public static RedisArray.Generic<long> CmsQuery(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            args.AddRange(items);
+            return new RedisArray.Generic<long>(new RedisInt("CMS.QUERY", args.ToArray()));
+        }
+        public static RedisStatus CmsMerge(string dest, long numKeys, string[] src, long[] weights)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { dest, numKeys });
+            args.AddRange(src);
+            if (weights?.Any() == true)
+            {
+                args.Add("WEIGHTS");
+                args.AddRange(weights.Select(a => (object)a));
+            }
+            return new RedisStatus("CMS.MERGE", args.ToArray());
+        }
+        public static RedisCmsInfoCommand CmsInfo(string key)
+        {
+            return new RedisCmsInfoCommand("CMS.INFO", key);
+        }
+        public class RedisCmsInfoCommand : RedisCommand<(long width, long depth, long count)>
+        {
+            public RedisCmsInfoCommand(string command, params object[] args)
+                : base(command, args)
+            {
+            }
+
+            public override (long width, long depth, long count) Parse(RedisReader reader)
+            {
+                long width = 0;
+                long depth = 0;
+                long count = 0;
+
+                reader.ExpectType(RedisMessage.MultiBulk);
+                long arrayCount = reader.ReadInt(false);
+                if (arrayCount % 2 != 0) throw new RedisProtocolException("CMS.INFO 返回数据格式 1级 MultiBulk 长度应该为偶数");
+
+                for (var b = 0; b < arrayCount; b += 2)
+                {
+                    var key = reader.ReadBulkString().ToLower();
+                    switch (key)
+                    {
+                        case "width":
+                            width = reader.ReadInt();
+                            break;
+                        case "depth":
+                            depth = reader.ReadInt();
+                            break;
+                        case "count":
+                            count = reader.ReadInt();
+                            break;
+                        default:
+                            reader.ReadBulkString();
+                            break;
+                    }
+                }
+                return (width, depth, count);
+            }
+        }
+        #endregion
+
+        #region RedisBloom TopK Filter
+        public static RedisStatus TopkReserve(string key, long topk, long width, long depth, decimal decay)
+        {
+            return new RedisStatus("TOPK.RESERVE", key, topk, width, depth, decay);
+        }
+        public static RedisArray.Generic<string> TopkAdd(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            args.AddRange(items);
+            return new RedisArray.Generic<string>(new RedisString("TOPK.ADD", args.ToArray()));
+        }
+        public static RedisArray.Generic<string> TopkIncrBy(string key, params (object item, long increment)[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            foreach (var item in items) args.AddRange(new object[] { item.item, item.increment });
+            return new RedisArray.Generic<string>(new RedisString("TOPK.INCRBY", args.ToArray()));
+        }
+        public static RedisArray.Generic<bool> TopkQuery(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            args.AddRange(items);
+            return new RedisArray.Generic<bool>(new RedisBool("TOPK.QUERY", args.ToArray()));
+        }
+        public static RedisArray.Generic<long> TopkCount(string key, object[] items)
+        {
+            var args = new List<object>();
+            args.AddRange(new object[] { key });
+            args.AddRange(items);
+            return new RedisArray.Generic<long>(new RedisInt("TOPK.COUNT", args.ToArray()));
+        }
+        public static RedisArray.Generic<string> TopkList(string key)
+        {
+            return new RedisArray.Generic<string>(new RedisString("TOPK.LIST", key));
+        }
+        public static RedisTopkInfoCommand TopkInfo(string key)
+        {
+            return new RedisTopkInfoCommand("TOPK.INFO", key);
+        }
+        public class RedisTopkInfoCommand : RedisCommand<(long k, long width, long depth, decimal decay)>
+        {
+            public RedisTopkInfoCommand(string command, params object[] args)
+                : base(command, args)
+            {
+            }
+
+            public override (long k, long width, long depth, decimal decay) Parse(RedisReader reader)
+            {
+                long k = 0;
+                long width = 0;
+                long depth = 0;
+                decimal decay = 0;
+
+                reader.ExpectType(RedisMessage.MultiBulk);
+                long arrayCount = reader.ReadInt(false);
+                if (arrayCount % 2 != 0) throw new RedisProtocolException("CMS.INFO 返回数据格式 1级 MultiBulk 长度应该为偶数");
+
+                for (var b = 0; b < arrayCount; b += 2)
+                {
+                    var key = reader.ReadBulkString().ToLower();
+                    switch (key)
+                    {
+                        case "k":
+                            k = reader.ReadInt();
+                            break;
+                        case "width":
+                            width = reader.ReadInt();
+                            break;
+                        case "depth":
+                            depth = reader.ReadInt();
+                            break;
+                        case "decay":
+                            decay = decimal.Parse(reader.ReadBulkString(), NumberStyles.Any);
+                            break;
+                        default:
+                            reader.ReadBulkString();
+                            break;
+                    }
+                }
+                return (k, width, depth, decay);
+            }
         }
         #endregion
 
@@ -1419,7 +2012,7 @@ namespace CSRedis
         }
     }
 
-    class RedisCommand
+    internal class RedisCommand
     {
         readonly string _command;
         readonly object[] _args;
@@ -1434,10 +2027,10 @@ namespace CSRedis
         }
     }
 
-    abstract class RedisCommand<T> : RedisCommand
+    internal abstract class RedisCommand<T> : RedisCommand
     {
         protected RedisCommand(string command, params object[] args)
-            : base (command,args)
+            : base(command, args)
         { }
 
         public abstract T Parse(RedisReader reader);
