@@ -1,9 +1,12 @@
 ﻿using CSRedis.Internal.Commands;
 using System;
 using System.Collections.Generic;
+using System.Collections.Concurrent;
 using System.Linq;
 using System.Threading.Tasks;
 
+#if net40
+#else
 namespace CSRedis
 {
     public partial class RedisClient
@@ -28,12 +31,23 @@ namespace CSRedis
             return WriteAsync(RedisCommands.Call(command, args));
         }
 
-        Task<T> WriteAsync<T>(RedisCommand<T> command)
+        internal ConcurrentQueue<TaskCompletionSource<object>> _asyncPipe;
+        async Task<T> WriteAsync<T>(RedisCommand<T> command)
         {
             if (_transaction.Active)
-                return _transaction.WriteAsync(command);
+                return await  _transaction.WriteAsync(command);
+            else if (_asyncPipe != null)
+            {
+                var tsc = new TaskCompletionSource<object>();
+                _asyncPipe.Enqueue(tsc);
+                
+                _connector.Pipeline.Write(command);
+
+                var ret = await tsc.Task;
+                return (T)ret;
+            }
             else
-                return _connector.CallAsync(command);
+                return await _connector.CallAsync(command);
         }
 
         #region Connection
@@ -93,6 +107,24 @@ namespace CSRedis
 
         #region Keys
         /// <summary>
+        /// [redis-server 3.2.1] 修改指定key(s) 最后访问时间 若key不存在，不做操作
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <returns></returns>
+        public Task<long> TouchAsync(params string[] keys)
+        {
+            return WriteAsync(RedisCommands.Touch(keys));
+        }
+        /// <summary>
+        /// [redis-server 4.0.0] Delete a key, 该命令和DEL十分相似：删除指定的key(s),若key不存在则该key被跳过。但是，相比DEL会产生阻塞，该命令会在另一个线程中回收内存，因此它是非阻塞的。 这也是该命令名字的由来：仅将keys从keyspace元数据中删除，真正的删除会在后续异步操作。
+        /// </summary>
+        /// <param name="keys">Keys to delete</param>
+        /// <returns>Number of keys removed</returns>
+        public Task<long> UnLinkAsync(params string[] keys)
+        {
+            return WriteAsync(RedisCommands.UnLink(keys));
+        }
+        /// <summary>
         /// Delete a key
         /// </summary>
         /// <param name="keys">Keys to delete</param>
@@ -120,6 +152,10 @@ namespace CSRedis
         public Task<bool> ExistsAsync(string key)
         {
             return WriteAsync(RedisCommands.Exists(key));
+        }
+        public Task<long> ExistsAsync(string[] keys)
+        {
+            return WriteAsync(RedisCommands.Exists(keys));
         }
 
         /// <summary>
@@ -396,29 +432,42 @@ namespace CSRedis
             return WriteAsync(RedisCommands.Type(key));
         }
 
-		/// <summary>
-		/// Iterate the set of keys in the currently selected Redis database
-		/// </summary>
-		/// <param name="cursor">The cursor returned by the server in the previous call, or 0 if this is the first call</param>
-		/// <param name="pattern">Glob-style pattern to filter returned elements</param>
-		/// <param name="count">Set the maximum number of elements to return</param>
-		/// <returns>Updated cursor and result set</returns>
-		public Task<RedisScan<string>> ScanAsync(long cursor, string pattern = null, long? count = null) {
-			return WriteAsync(RedisCommands.Scan(cursor, pattern, count));
-		}
-		public Task<RedisScan<byte[]>> ScanBytesAsync(long cursor, string pattern = null, long? count = null) {
-			return WriteAsync(RedisCommands.ScanBytes(cursor, pattern, count));
-		}
-		#endregion
+        /// <summary>
+        /// Iterate the set of keys in the currently selected Redis database
+        /// </summary>
+        /// <param name="cursor">The cursor returned by the server in the previous call, or 0 if this is the first call</param>
+        /// <param name="pattern">Glob-style pattern to filter returned elements</param>
+        /// <param name="count">Set the maximum number of elements to return</param>
+        /// <returns>Updated cursor and result set</returns>
+        public Task<RedisScan<string>> ScanAsync(long cursor, string pattern = null, long? count = null)
+        {
+            return WriteAsync(RedisCommands.Scan(cursor, pattern, count));
+        }
+        public Task<RedisScan<byte[]>> ScanBytesAsync(long cursor, string pattern = null, long? count = null)
+        {
+            return WriteAsync(RedisCommands.ScanBytes(cursor, pattern, count));
+        }
+        #endregion
 
-		#region Hashes
-		/// <summary>
-		/// Delete one or more hash fields
-		/// </summary>
-		/// <param name="key">Hash key</param>
-		/// <param name="fields">Fields to delete</param>
-		/// <returns>Number of fields removed from hash</returns>
-		public Task<long> HDelAsync(string key, params string[] fields)
+        #region Hashes
+        /// <summary>
+        /// [redis-server 3.2.0] 返回hash指定field的value的字符串长度，如果hash或者field不存在，返回0.
+        /// </summary>
+        /// <param name="key">Hash key</param>
+        /// <param name="field">Field</param>
+        /// <returns></returns>
+        public Task<long> HStrLenAsync(string key, string field)
+        {
+            return WriteAsync(RedisCommands.HStrLen(key, field));
+        }
+
+        /// <summary>
+        /// Delete one or more hash fields
+        /// </summary>
+        /// <param name="key">Hash key</param>
+        /// <param name="fields">Fields to delete</param>
+        /// <returns>Number of fields removed from hash</returns>
+        public Task<long> HDelAsync(string key, params string[] fields)
         {
             return WriteAsync(RedisCommands.HDel(key, fields));
         }
@@ -444,7 +493,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.HGet(key, field));
         }
-		public Task<byte[]> HGetBytesAsync(string key, string field)
+        public Task<byte[]> HGetBytesAsync(string key, string field)
         {
             return WriteAsync(RedisCommands.HGetBytes(key, field));
         }
@@ -470,7 +519,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.HGetAll(key));
         }
-		public Task<Dictionary<string, byte[]>> HGetAllBytesAsync(string key)
+        public Task<Dictionary<string, byte[]>> HGetAllBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.HGetAllBytes(key));
         }
@@ -494,7 +543,7 @@ namespace CSRedis
         /// <param name="field">Field to increment</param>
         /// <param name="increment">Increment value</param>
         /// <returns>Value of field after increment</returns>
-        public Task<double> HIncrByFloatAsync(string key, string field, double increment)
+        public Task<decimal> HIncrByFloatAsync(string key, string field, decimal increment)
         {
             return WriteAsync(RedisCommands.HIncrByFloat(key, field, increment));
         }
@@ -529,7 +578,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.HMGet(key, fields));
         }
-		public Task<byte[][]> HMGetBytesAsync(string key, params string[] fields)
+        public Task<byte[][]> HMGetBytesAsync(string key, params string[] fields)
         {
             return WriteAsync(RedisCommands.HMGetBytes(key, fields));
         }
@@ -602,12 +651,12 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.HVals(key));
         }
-		public Task<byte[][]> HValsBytesAsync(string key)
+        public Task<byte[][]> HValsBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.HValsBytes(key));
         }
 
-		/// <summary>
+        /// <summary>
         /// Iterate the keys and values of a hash field
         /// </summary>
         /// <param name="key">Hash key</param>
@@ -619,24 +668,24 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.HScan(key, cursor, pattern, count));
         }
-		public Task<RedisScan<Tuple<string, byte[]>>> HScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
+        public Task<RedisScan<Tuple<string, byte[]>>> HScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
         {
             return WriteAsync(RedisCommands.HScanBytes(key, cursor, pattern, count));
         }
-		#endregion
+        #endregion
 
-		#region Lists
-		/// <summary>
-		/// Get an element from a list by its index
-		/// </summary>
-		/// <param name="key">List key</param>
-		/// <param name="index">Zero-based index of item to return</param>
-		/// <returns>Element at index</returns>
-		public Task<string> LIndexAsync(string key, long index)
+        #region Lists
+        /// <summary>
+        /// Get an element from a list by its index
+        /// </summary>
+        /// <param name="key">List key</param>
+        /// <param name="index">Zero-based index of item to return</param>
+        /// <returns>Element at index</returns>
+        public Task<string> LIndexAsync(string key, long index)
         {
             return WriteAsync(RedisCommands.LIndex(key, index));
         }
-		public Task<byte[]> LIndexBytesAsync(string key, long index)
+        public Task<byte[]> LIndexBytesAsync(string key, long index)
         {
             return WriteAsync(RedisCommands.LIndexBytes(key, index));
         }
@@ -673,7 +722,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.LPop(key));
         }
-		public Task<byte[]> LPopBytesAsync(string key)
+        public Task<byte[]> LPopBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.LPopBytes(key));
         }
@@ -711,7 +760,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.LRange(key, start, stop));
         }
-		public Task<byte[][]> LRangeBytesAsync(string key, long start, long stop)
+        public Task<byte[][]> LRangeBytesAsync(string key, long start, long stop)
         {
             return WriteAsync(RedisCommands.LRangeBytes(key, start, stop));
         }
@@ -761,7 +810,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.RPop(key));
         }
-		public Task<byte[]> RPopBytesAsync(string key)
+        public Task<byte[]> RPopBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.RPopBytes(key));
         }
@@ -776,7 +825,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.RPopLPush(source, destination));
         }
-		public Task<byte[]> RPopBytesLPushAsync(string source, string destination)
+        public Task<byte[]> RPopBytesLPushAsync(string source, string destination)
         {
             return WriteAsync(RedisCommands.RPopBytesLPush(source, destination));
         }
@@ -835,7 +884,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SDiff(keys));
         }
-		public Task<byte[][]> SDiffBytesAsync(params string[] keys)
+        public Task<byte[][]> SDiffBytesAsync(params string[] keys)
         {
             return WriteAsync(RedisCommands.SDiffBytes(keys));
         }
@@ -860,7 +909,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SInter(keys));
         }
-		public Task<byte[][]> SInterBytesAsync(params string[] keys)
+        public Task<byte[][]> SInterBytesAsync(params string[] keys)
         {
             return WriteAsync(RedisCommands.SInterBytes(keys));
         }
@@ -896,7 +945,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SMembers(key));
         }
-		public Task<byte[][]> SMembersBytesAsync(string key)
+        public Task<byte[][]> SMembersBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.SMembersBytes(key));
         }
@@ -922,9 +971,24 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SPop(key));
         }
-		public Task<byte[]> SPopBytesAsync(string key)
+        public Task<byte[]> SPopBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.SPopBytes(key));
+        }
+
+        /// <summary>
+        /// Remove and return one or more random members from a set
+        /// </summary>
+        /// <param name="key">Set key</param>
+        /// <param name="count">Number of elements to remove and return</param>
+        /// <returns></returns>
+        public Task<string[]> SPopAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.SPop(key, count));
+        }
+        public Task<byte[][]> SPopBytesAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.SPopBytes(key, count));
         }
 
         /// <summary>
@@ -936,7 +1000,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SRandMember(key));
         }
-		public Task<byte[]> SRandMemberBytesAsync(string key)
+        public Task<byte[]> SRandMemberBytesAsync(string key)
         {
             return WriteAsync(RedisCommands.SRandMemberBytes(key));
         }
@@ -951,7 +1015,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SRandMembers(key, count));
         }
-		public Task<byte[][]> SRandMembersBytesAsync(string key, long count)
+        public Task<byte[][]> SRandMembersBytesAsync(string key, long count)
         {
             return WriteAsync(RedisCommands.SRandMembersBytes(key, count));
         }
@@ -976,7 +1040,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SUnion(keys));
         }
-		public Task<byte[][]> SUnionBytesAsync(params string[] keys)
+        public Task<byte[][]> SUnionBytesAsync(params string[] keys)
         {
             return WriteAsync(RedisCommands.SUnionBytes(keys));
         }
@@ -992,7 +1056,7 @@ namespace CSRedis
             return WriteAsync(RedisCommands.SUnionStore(destination, keys));
         }
 
-		/// <summary>
+        /// <summary>
         /// Iterate the elements of a set field
         /// </summary>
         /// <param name="key">Set key</param>
@@ -1004,13 +1068,31 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.SScan(key, cursor, pattern, count));
         }
-		public Task<RedisScan<byte[]>> SScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
+        public Task<RedisScan<byte[]>> SScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
         {
             return WriteAsync(RedisCommands.SScanBytes(key, cursor, pattern, count));
         }
         #endregion
 
         #region Sorted Sets
+        public Task<Tuple<string, decimal>[]> ZPopMaxAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.ZPopMax(key, count));
+        }
+        public Task<Tuple<byte[], decimal>[]> ZPopMaxBytesAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.ZPopMaxBytes(key, count));
+        }
+        public Task<Tuple<string, decimal>[]> ZPopMinAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.ZPopMin(key, count));
+        }
+        public Task<Tuple<byte[], decimal>[]> ZPopMinBytesAsync(string key, long count)
+        {
+            return WriteAsync(RedisCommands.ZPopMinBytes(key, count));
+        }
+
+
         /// <summary>
         /// Add one or more members to a sorted set, or update its score if it already exists
         /// </summary>
@@ -1052,7 +1134,7 @@ namespace CSRedis
         /// <param name="exclusiveMin">Minimum score is exclusive</param>
         /// <param name="exclusiveMax">Maximum score is exclusive</param>
         /// <returns>Number of elements in the specified score range</returns>
-        public Task<long> ZCountAsync(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)
+        public Task<long> ZCountAsync(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             return WriteAsync(RedisCommands.ZCount(key, min, max, exclusiveMin, exclusiveMax));
         }
@@ -1076,7 +1158,7 @@ namespace CSRedis
         /// <param name="increment">Increment by value</param>
         /// <param name="member">Sorted set member to increment</param>
         /// <returns>New score of member</returns>
-        public Task<double> ZIncrByAsync(string key, double increment, string member)
+        public Task<decimal> ZIncrByAsync(string key, decimal increment, object member)
         {
             return WriteAsync(RedisCommands.ZIncrBy(key, increment, member));
         }
@@ -1089,7 +1171,7 @@ namespace CSRedis
         /// <param name="aggregate">Aggregation function of resulting set</param>
         /// <param name="keys">Sorted set keys to intersect</param>
         /// <returns>Number of elements in the resulting sorted set</returns>
-        public Task<long> ZInterStoreAsync(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public Task<long> ZInterStoreAsync(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             return WriteAsync(RedisCommands.ZInterStore(destination, weights, aggregate, keys));
         }
@@ -1117,7 +1199,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.ZRange(key, start, stop, withScores));
         }
-		public Task<byte[][]> ZRangeBytesAsync(string key, long start, long stop, bool withScores = false)
+        public Task<byte[][]> ZRangeBytesAsync(string key, long start, long stop, bool withScores = false)
         {
             return WriteAsync(RedisCommands.ZRangeBytes(key, start, stop, withScores));
         }
@@ -1129,11 +1211,11 @@ namespace CSRedis
         /// <param name="start">Start offset</param>
         /// <param name="stop">Stop offset</param>
         /// <returns>Array of elements in the specified range with scores</returns>
-        public Task<Tuple<string, double>[]> ZRangeWithScoresAsync(string key, long start, long stop)
+        public Task<Tuple<string, decimal>[]> ZRangeWithScoresAsync(string key, long start, long stop)
         {
             return WriteAsync(RedisCommands.ZRangeWithScores(key, start, stop));
         }
-		public Task<Tuple<byte[], double>[]> ZRangeBytesWithScoresAsync(string key, long start, long stop)
+        public Task<Tuple<byte[], decimal>[]> ZRangeBytesWithScoresAsync(string key, long start, long stop)
         {
             return WriteAsync(RedisCommands.ZRangeBytesWithScores(key, start, stop));
         }
@@ -1150,11 +1232,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Task<string[]> ZRangeByScoreAsync(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Task<string[]> ZRangeByScoreAsync(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeByScore(key, min, max, withScores, exclusiveMin, exclusiveMax, offset, count));
         }
-		public Task<byte[][]> ZRangeBytesByScoreAsync(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Task<byte[][]> ZRangeBytesByScoreAsync(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeBytesByScore(key, min, max, withScores, exclusiveMin, exclusiveMax, offset, count));
         }
@@ -1173,7 +1255,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.ZRangeByScore(key, min, max, withScores, offset, count));
         }
-		public Task<byte[][]> ZRangeBytesByScoreAsync(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
+        public Task<byte[][]> ZRangeBytesByScoreAsync(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeBytesByScore(key, min, max, withScores, offset, count));
         }
@@ -1189,11 +1271,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Task<Tuple<string, double>[]> ZRangeByScoreWithScoresAsync(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Task<Tuple<string, decimal>[]> ZRangeByScoreWithScoresAsync(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeByScoreWithScores(key, min, max, exclusiveMin, exclusiveMax, offset, count));
         }
-		public Task<Tuple<byte[], double>[]> ZRangeBytesByScoreWithScoresAsync(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Task<Tuple<byte[], decimal>[]> ZRangeBytesByScoreWithScoresAsync(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeBytesByScoreWithScores(key, min, max, exclusiveMin, exclusiveMax, offset, count));
         }
@@ -1207,11 +1289,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Task<Tuple<string, double>[]> ZRangeByScoreWithScoresAsync(string key, string min, string max, long? offset = null, long? count = null)
+        public Task<Tuple<string, decimal>[]> ZRangeByScoreWithScoresAsync(string key, string min, string max, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeByScoreWithScores(key, min, max, offset, count));
         }
-		public Task<Tuple<byte[], double>[]> ZRangeBytesByScoreWithScoresAsync(string key, string min, string max, long? offset = null, long? count = null)
+        public Task<Tuple<byte[], decimal>[]> ZRangeBytesByScoreWithScoresAsync(string key, string min, string max, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeBytesByScoreWithScores(key, min, max, offset, count));
         }
@@ -1259,11 +1341,11 @@ namespace CSRedis
         /// <param name="exclusiveMin">Minimum score is exclusive</param>
         /// <param name="exclusiveMax">Maximum score is exclusive</param>
         /// <returns>Number of elements removed</returns>
-        public Task<long> ZRemRangeByScoreAsync(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)
+        public Task<long> ZRemRangeByScoreAsync(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             return WriteAsync(RedisCommands.ZRemRangeByScore(key, min, max, exclusiveMin, exclusiveMax));
         }
-		public Task<long> ZRemRangeByScoreAsync(string key, string min, string max)
+        public Task<long> ZRemRangeByScoreAsync(string key, string min, string max)
         {
             return WriteAsync(RedisCommands.ZRemRangeByScore(key, min, max));
         }
@@ -1280,7 +1362,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.ZRevRange(key, start, stop, withScores));
         }
-		public Task<byte[][]> ZRevRangeBytesAsync(string key, long start, long stop, bool withScores = false)
+        public Task<byte[][]> ZRevRangeBytesAsync(string key, long start, long stop, bool withScores = false)
         {
             return WriteAsync(RedisCommands.ZRevRangeBytes(key, start, stop, withScores));
         }
@@ -1292,11 +1374,11 @@ namespace CSRedis
         /// <param name="start">Start offset</param>
         /// <param name="stop">Stop offset</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Task<Tuple<string, double>[]> ZRevRangeWithScoresAsync(string key, long start, long stop)
+        public Task<Tuple<string, decimal>[]> ZRevRangeWithScoresAsync(string key, long start, long stop)
         {
             return WriteAsync(RedisCommands.ZRevRangeWithScores(key, start, stop));
         }
-		public Task<Tuple<byte[], double>[]> ZRevRangeBytesWithScoresAsync(string key, long start, long stop)
+        public Task<Tuple<byte[], decimal>[]> ZRevRangeBytesWithScoresAsync(string key, long start, long stop)
         {
             return WriteAsync(RedisCommands.ZRevRangeBytesWithScores(key, start, stop));
         }
@@ -1313,29 +1395,30 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public Task<string[]> ZRevRangeByScoreAsync(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public Task<string[]> ZRevRangeByScoreAsync(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeByScore(key, max, min, withScores, exclusiveMax, exclusiveMin, offset, count));
         }
-		public Task<byte[][]> ZRevRangeBytesByScoreAsync(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null) {
-			return WriteAsync(RedisCommands.ZRevRangeBytesByScore(key, max, min, withScores, exclusiveMax, exclusiveMin, offset, count));
-		}
+        public Task<byte[][]> ZRevRangeBytesByScoreAsync(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        {
+            return WriteAsync(RedisCommands.ZRevRangeBytesByScore(key, max, min, withScores, exclusiveMax, exclusiveMin, offset, count));
+        }
 
-		/// <summary>
-		/// Return a range of members in a sorted set, by score, with scores ordered from high to low
-		/// </summary>
-		/// <param name="key">Sorted set key</param>
-		/// <param name="max">Maximum score</param>
-		/// <param name="min">Minimum score</param>
-		/// <param name="withScores">Include scores in result</param>
-		/// <param name="offset">Start offset</param>
-		/// <param name="count">Number of elements to return</param>
-		/// <returns>List of elements in the specified score range (with optional scores)</returns>
-		public Task<string[]> ZRevRangeByScoreAsync(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
+        /// <summary>
+        /// Return a range of members in a sorted set, by score, with scores ordered from high to low
+        /// </summary>
+        /// <param name="key">Sorted set key</param>
+        /// <param name="max">Maximum score</param>
+        /// <param name="min">Minimum score</param>
+        /// <param name="withScores">Include scores in result</param>
+        /// <param name="offset">Start offset</param>
+        /// <param name="count">Number of elements to return</param>
+        /// <returns>List of elements in the specified score range (with optional scores)</returns>
+        public Task<string[]> ZRevRangeByScoreAsync(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeByScore(key, max, min, withScores, offset, count));
         }
-		public Task<byte[][]> ZRevRangeBytesByScoreAsync(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
+        public Task<byte[][]> ZRevRangeBytesByScoreAsync(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeBytesByScore(key, max, min, withScores, offset, count));
         }
@@ -1351,11 +1434,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public Task<Tuple<string, double>[]> ZRevRangeByScoreWithScoresAsync(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public Task<Tuple<string, decimal>[]> ZRevRangeByScoreWithScoresAsync(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeByScoreWithScores(key, max, min, exclusiveMax, exclusiveMin, offset, count));
         }
-		public Task<Tuple<byte[], double>[]> ZRevRangeBytesByScoreWithScoresAsync(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public Task<Tuple<byte[], decimal>[]> ZRevRangeBytesByScoreWithScoresAsync(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeBytesByScoreWithScores(key, max, min, exclusiveMax, exclusiveMin, offset, count));
         }
@@ -1369,11 +1452,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public Task<Tuple<string, double>[]> ZRevRangeByScoreWithScoresAsync(string key, string max, string min, long? offset = null, long? count = null)
+        public Task<Tuple<string, decimal>[]> ZRevRangeByScoreWithScoresAsync(string key, string max, string min, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeByScoreWithScores(key, max, min, offset, count));
         }
-		public Task<Tuple<byte[], double>[]> ZRevRangeBytesByScoreWithScoresAsync(string key, string max, string min, long? offset = null, long? count = null)
+        public Task<Tuple<byte[], decimal>[]> ZRevRangeBytesByScoreWithScoresAsync(string key, string max, string min, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRevRangeBytesByScoreWithScores(key, max, min, offset, count));
         }
@@ -1395,7 +1478,7 @@ namespace CSRedis
         /// <param name="key">Sorted set key</param>
         /// <param name="member">Member to lookup</param>
         /// <returns>Score of member, or null if member does not exist</returns>
-        public Task<double?> ZScoreAsync(string key, object member)
+        public Task<decimal?> ZScoreAsync(string key, object member)
         {
             return WriteAsync(RedisCommands.ZScore(key, member));
         }
@@ -1408,7 +1491,7 @@ namespace CSRedis
         /// <param name="aggregate">Aggregation function of resulting set</param>
         /// <param name="keys">Sorted set keys to union</param>
         /// <returns>Number of elements in the resulting sorted set</returns>
-        public Task<long> ZUnionStoreAsync(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public Task<long> ZUnionStoreAsync(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             return WriteAsync(RedisCommands.ZUnionStore(destination, weights, aggregate, keys));
         }
@@ -1421,11 +1504,11 @@ namespace CSRedis
         /// <param name="pattern">Glob-style pattern to filter returned elements</param>
         /// <param name="count">Maximum number of elements to return</param>
         /// <returns>Updated cursor and result set</returns>
-        public Task<RedisScan<Tuple<string, double>>> ZScanAsync(string key, long cursor, string pattern = null, long? count = null)
+        public Task<RedisScan<Tuple<string, decimal>>> ZScanAsync(string key, long cursor, string pattern = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZScan(key, cursor, pattern, count));
         }
-		public Task<RedisScan<Tuple<byte[], double>>> ZScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
+        public Task<RedisScan<Tuple<byte[], decimal>>> ZScanBytesAsync(string key, long cursor, string pattern = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZScanBytes(key, cursor, pattern, count));
         }
@@ -1443,7 +1526,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.ZRangeByLex(key, min, max, offset, count));
         }
-		public Task<byte[][]> ZRangeBytesByLexAsync(string key, string min, string max, long? offset = null, long? count = null)
+        public Task<byte[][]> ZRangeBytesByLexAsync(string key, string min, string max, long? offset = null, long? count = null)
         {
             return WriteAsync(RedisCommands.ZRangeBytesByLex(key, min, max, offset, count));
         }
@@ -1540,12 +1623,12 @@ namespace CSRedis
             return WriteAsync(RedisCommands.EvalSHA(sha1, keys, arguments));
         }
 
-		/// <summary>
-		/// Check existence of script SHA hashes in the script cache
-		/// </summary>
-		/// <param name="sha1s">SHA1 script hashes</param>
-		/// <returns>Array of boolean values indicating script existence on server</returns>
-		public Task<bool[]> ScriptExistsAsync(params string[] sha1s)
+        /// <summary>
+        /// Check existence of script SHA hashes in the script cache
+        /// </summary>
+        /// <param name="sha1s">SHA1 script hashes</param>
+        /// <returns>Array of boolean values indicating script existence on server</returns>
+        public Task<bool[]> ScriptExistsAsync(params string[] sha1s)
         {
             return WriteAsync(RedisCommands.ScriptExists(sha1s));
         }
@@ -1658,18 +1741,18 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.Get(key));
         }
-		public Task<byte[]> GetBytesAsync(string key)
-		{
-			return WriteAsync(RedisCommands.GetBytes(key));
-		}
+        public Task<byte[]> GetBytesAsync(string key)
+        {
+            return WriteAsync(RedisCommands.GetBytes(key));
+        }
 
-		/// <summary>
-		/// Returns the bit value at offset in the string value stored at key
-		/// </summary>
-		/// <param name="key">Key to lookup</param>
-		/// <param name="offset">Offset of key to check</param>
-		/// <returns>Bit value stored at offset</returns>
-		public Task<bool> GetBitAsync(string key, uint offset)
+        /// <summary>
+        /// Returns the bit value at offset in the string value stored at key
+        /// </summary>
+        /// <param name="key">Key to lookup</param>
+        /// <param name="offset">Offset of key to check</param>
+        /// <returns>Bit value stored at offset</returns>
+        public Task<bool> GetBitAsync(string key, uint offset)
         {
             return WriteAsync(RedisCommands.GetBit(key, offset));
         }
@@ -1685,7 +1768,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.GetRange(key, start, end));
         }
-		public Task<byte[]> GetRangeBytesAsync(string key, long start, long end)
+        public Task<byte[]> GetRangeBytesAsync(string key, long start, long end)
         {
             return WriteAsync(RedisCommands.GetRangeBytes(key, start, end));
         }
@@ -1700,7 +1783,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.GetSet(key, value));
         }
-		public Task<byte[]> GetSetBytesAsync(string key, object value)
+        public Task<byte[]> GetSetBytesAsync(string key, object value)
         {
             return WriteAsync(RedisCommands.GetSetBytes(key, value));
         }
@@ -1732,7 +1815,7 @@ namespace CSRedis
         /// <param name="key">Key to modify</param>
         /// <param name="increment">Increment amount</param>
         /// <returns>Value of key after increment</returns>
-        public Task<double> IncrByFloatAsync(string key, double increment)
+        public Task<decimal> IncrByFloatAsync(string key, decimal increment)
         {
             return WriteAsync(RedisCommands.IncrByFloat(key, increment));
         }
@@ -1746,7 +1829,7 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.MGet(keys));
         }
-		public Task<byte[][]> MGetBytesAsync(params string[] keys)
+        public Task<byte[][]> MGetBytesAsync(params string[] keys)
         {
             return WriteAsync(RedisCommands.MGetBytes(keys));
         }
@@ -2269,99 +2352,108 @@ namespace CSRedis
         {
             return WriteAsync(RedisCommands.PfMerge(destKey, sourceKeys));
         }
-		#endregion
+        #endregion
 
-		#region Geo redis-server 3.2
-		public Task<long> GeoAddAsync(string key, params (double longitude, double latitude, object member)[] values) {
-			if (values == null || values.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			foreach (var v in values) args.AddRange(new object[] { v.longitude, v.latitude, v.member });
-			return WriteAsync(new RedisInt("GEOADD", args.ToArray()));
-		}
-		public Task<double?> GeoDistAsync(string key, object member1, object member2, GeoUnit unit = GeoUnit.m) {
-			if (unit == GeoUnit.m) return WriteAsync(new RedisFloat.Nullable("GEODIST", key, member1, member2));
-			return WriteAsync(new RedisFloat.Nullable("GEODIST", key, member1, member2, unit));
-		}
-		public Task<string[]> GeoHashAsync(string key, object[] members) {
-			if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			args.AddRange(members);
-			return WriteAsync(new RedisArray.Strings("GEOHASH", args.ToArray()));
-		}
-		async public Task<(double longitude, double latitude)?[]> GeoPosAsync(string key, object[] members) {
-			if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			args.AddRange(members);
-			var ret = await WriteAsync(new RedisArray.Generic<double[]>(new RedisArray.Generic<double>(new RedisFloat("GEOPOS", args.ToArray()))));
-			return ret.Select(a => a != null && a.Length == 2 ? new(double, double)?((a[0], a[1])) : null).ToArray();
-		}
-		async public Task<(string member, double dist, double longitude, double latitude, long hash)[]> GeoRadiusAsync(string key, double longitude, double latitude, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+        #region Geo redis-server 3.2
+        public Task<long> GeoAddAsync(string key, params (decimal longitude, decimal latitude, object member)[] values)
+        {
+            if (values == null || values.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            foreach (var v in values) args.AddRange(new object[] { v.longitude, v.latitude, v.member });
+            return WriteAsync(new RedisInt("GEOADD", args.ToArray()));
+        }
+        public Task<decimal?> GeoDistAsync(string key, object member1, object member2, GeoUnit unit = GeoUnit.m)
+        {
+            if (unit == GeoUnit.m) return WriteAsync(new RedisFloat.Nullable("GEODIST", key, member1, member2));
+            return WriteAsync(new RedisFloat.Nullable("GEODIST", key, member1, member2, unit));
+        }
+        public Task<string[]> GeoHashAsync(string key, object[] members)
+        {
+            if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(members);
+            return WriteAsync(new RedisArray.Strings("GEOHASH", args.ToArray()));
+        }
+        async public Task<(decimal longitude, decimal latitude)?[]> GeoPosAsync(string key, object[] members)
+        {
+            if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(members);
+            var ret = await WriteAsync(new RedisArray.Generic<decimal[]>(new RedisArray.Generic<decimal>(new RedisFloat("GEOPOS", args.ToArray()))));
+            return ret.Select(a => a != null && a.Length == 2 ? new (decimal, decimal)?((a[0], a[1])) : null).ToArray();
+        }
+        async public Task<(string member, decimal dist, decimal longitude, decimal latitude, long hash)[]> GeoRadiusAsync(string key, decimal longitude, decimal latitude, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.Add(count);
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<string, double, long, double[]>.Single(
-				new RedisString(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
-			var ret = await WriteAsync(new RedisArray.Generic<Tuple<string, double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		async public Task<(byte[] member, double dist, double longitude, double latitude, long hash)[]> GeoRadiusBytesAsync(string key, double longitude, double latitude, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<string, decimal, long, decimal[]>.Single(
+                new RedisString(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
+            var ret = await WriteAsync(new RedisArray.Generic<Tuple<string, decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        async public Task<(byte[] member, decimal dist, decimal longitude, decimal latitude, long hash)[]> GeoRadiusBytesAsync(string key, decimal longitude, decimal latitude, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.Add(count);
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<byte[], double, long, double[]>.Single(
-				new RedisBytes(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
-			var ret = await WriteAsync(new RedisArray.Generic<Tuple<byte[], double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		async public Task<(string member, double dist, double longitude, double latitude, long hash)[]> GeoRadiusByMemberAsync(string key, object member, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, member, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<byte[], decimal, long, decimal[]>.Single(
+                new RedisBytes(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
+            var ret = await WriteAsync(new RedisArray.Generic<Tuple<byte[], decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        async public Task<(string member, decimal dist, decimal longitude, decimal latitude, long hash)[]> GeoRadiusByMemberAsync(string key, object member, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, member, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.Add(count);
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<string, double, long, double[]>.Single(
-				new RedisString(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
-			var ret = await WriteAsync(new RedisArray.Generic<Tuple<string, double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		async public Task<(byte[] member, double dist, double longitude, double latitude, long hash)[]> GeoRadiusBytesByMemberAsync(string key, object member, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, member, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<string, decimal, long, decimal[]>.Single(
+                new RedisString(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
+            var ret = await WriteAsync(new RedisArray.Generic<Tuple<string, decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        async public Task<(byte[] member, decimal dist, decimal longitude, decimal latitude, long hash)[]> GeoRadiusBytesByMemberAsync(string key, object member, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, member, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.Add(count);
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<byte[], double, long, double[]>.Single(
-				new RedisBytes(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
-			var ret = await WriteAsync(new RedisArray.Generic<Tuple<byte[], double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		#endregion
-	}
+            var cmd = new RedisTuple.Generic<byte[], decimal, long, decimal[]>.Single(
+                new RedisBytes(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
+            var ret = await WriteAsync(new RedisArray.Generic<Tuple<byte[], decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        #endregion
+    }
 }
+#endif
