@@ -28,7 +28,7 @@ namespace CSRedis
             return Write(RedisCommands.Call(command, args));
         }
 
-        T Write<T>(RedisCommand<T> command)
+        internal T Write<T>(RedisCommand<T> command)
         {
             if (_transaction.Active)
                 return _transaction.Write(command);
@@ -42,6 +42,8 @@ namespace CSRedis
             else
                 return _connector.Call(command);
         }
+
+        internal void WriteNoneRead(RedisCommand command) => _connector.CallNoneRead(command);
 
         #region Connection
         /// <summary>
@@ -97,6 +99,24 @@ namespace CSRedis
 
         #region Keys
         /// <summary>
+        /// [redis-server 3.2.1] 修改指定key(s) 最后访问时间 若key不存在，不做操作
+        /// </summary>
+        /// <param name="keys">Keys</param>
+        /// <returns></returns>
+        public long Touch(params string[] keys)
+        {
+            return Write(RedisCommands.Touch(keys));
+        }
+        /// <summary>
+        /// [redis-server 4.0.0] Delete a key, 该命令和DEL十分相似：删除指定的key(s),若key不存在则该key被跳过。但是，相比DEL会产生阻塞，该命令会在另一个线程中回收内存，因此它是非阻塞的。 这也是该命令名字的由来：仅将keys从keyspace元数据中删除，真正的删除会在后续异步操作。
+        /// </summary>
+        /// <param name="keys">Keys to delete</param>
+        /// <returns>Number of keys removed</returns>
+        public long UnLink(params string[] keys)
+        {
+            return Write(RedisCommands.UnLink(keys));
+        }
+        /// <summary>
         /// Delete a key
         /// </summary>
         /// <param name="keys">Keys to delete</param>
@@ -124,6 +144,10 @@ namespace CSRedis
         public bool Exists(string key)
         {
             return Write(RedisCommands.Exists(key));
+        }
+        public long Exists(string[] keys)
+        {
+            return Write(RedisCommands.Exists(keys));
         }
 
         /// <summary>
@@ -335,14 +359,14 @@ namespace CSRedis
             return Write(RedisCommands.RenameNx(key, newKey));
         }
 
-		/// <summary>
-		/// Create a key using the provided serialized value, previously obtained using dump
-		/// </summary>
-		/// <param name="key">Key to restore</param>
-		/// <param name="ttlMilliseconds">Time-to-live in milliseconds</param>
-		/// <param name="serializedValue">Serialized value from DUMP</param>
-		/// <returns>Status code</returns>
-		public string Restore(string key, long ttlMilliseconds, byte[] serializedValue)
+        /// <summary>
+        /// Create a key using the provided serialized value, previously obtained using dump
+        /// </summary>
+        /// <param name="key">Key to restore</param>
+        /// <param name="ttlMilliseconds">Time-to-live in milliseconds</param>
+        /// <param name="serializedValue">Serialized value from DUMP</param>
+        /// <returns>Status code</returns>
+        public string Restore(string key, long ttlMilliseconds, byte[] serializedValue)
         {
             return Write(RedisCommands.Restore(key, ttlMilliseconds, serializedValue));
         }
@@ -411,13 +435,24 @@ namespace CSRedis
         {
             return Write(RedisCommands.Scan(cursor, pattern, count));
         }
-		public RedisScan<byte[]> ScanBytes(long cursor, string pattern = null, long? count = null)
+        public RedisScan<byte[]> ScanBytes(long cursor, string pattern = null, long? count = null)
         {
             return Write(RedisCommands.ScanBytes(cursor, pattern, count));
         }
         #endregion
 
         #region Hashes
+        /// <summary>
+        /// [redis-server 3.2.0] 返回hash指定field的value的字符串长度，如果hash或者field不存在，返回0.
+        /// </summary>
+        /// <param name="key">Hash key</param>
+        /// <param name="field">Field</param>
+        /// <returns></returns>
+        public long HStrLen(string key, string field)
+        {
+            return Write(RedisCommands.HStrLen(key, field));
+        }
+
         /// <summary>
         /// Delete one or more hash fields
         /// </summary>
@@ -448,12 +483,12 @@ namespace CSRedis
         /// <returns>Value of hash field</returns>
         public string HGet(string key, string field)
         {
-			return Write(RedisCommands.HGet(key, field));
-		}
-		public byte[] HGetBytes(string key, string field)
+            return Write(RedisCommands.HGet(key, field));
+        }
+        public byte[] HGetBytes(string key, string field)
         {
-			return Write(RedisCommands.HGetBytes(key, field));
-		}
+            return Write(RedisCommands.HGetBytes(key, field));
+        }
 
         /// <summary>
         /// Get all the fields and values in a hash
@@ -476,7 +511,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.HGetAll(key));
         }
-		public Dictionary<string, byte[]> HGetAllBytes(string key)
+        public Dictionary<string, byte[]> HGetAllBytes(string key)
         {
             return Write(RedisCommands.HGetAllBytes(key));
         }
@@ -500,7 +535,7 @@ namespace CSRedis
         /// <param name="field">Field to increment</param>
         /// <param name="increment">Increment value</param>
         /// <returns>Value of field after increment</returns>
-        public double HIncrByFloat(string key, string field, double increment)
+        public decimal HIncrByFloat(string key, string field, decimal increment)
         {
             return Write(RedisCommands.HIncrByFloat(key, field, increment));
         }
@@ -535,7 +570,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.HMGet(key, fields));
         }
-		public byte[][] HMGetBytes(string key, params string[] fields)
+        public byte[][] HMGetBytes(string key, params string[] fields)
         {
             return Write(RedisCommands.HMGetBytes(key, fields));
         }
@@ -608,7 +643,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.HVals(key));
         }
-		public byte[][] HValsBytes(string key)
+        public byte[][] HValsBytes(string key)
         {
             return Write(RedisCommands.HValsBytes(key));
         }
@@ -625,7 +660,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.HScan(key, cursor, pattern, count));
         }
-		public RedisScan<Tuple<string, byte[]>> HScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        public RedisScan<Tuple<string, byte[]>> HScanBytes(string key, long cursor, string pattern = null, long? count = null)
         {
             return Write(RedisCommands.HScanBytes(key, cursor, pattern, count));
         }
@@ -642,7 +677,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BLPopWithKey(timeout, keys));
         }
-		public Tuple<string, byte[]> BLPopBytesWithKey(int timeout, params string[] keys)
+        public Tuple<string, byte[]> BLPopBytesWithKey(int timeout, params string[] keys)
         {
             return Write(RedisCommands.BLPopBytesWithKey(timeout, keys));
         }
@@ -657,7 +692,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BLPopWithKey(timeout, keys));
         }
-		public Tuple<string, byte[]> BLPopBytesWithKey(TimeSpan timeout, params string[] keys)
+        public Tuple<string, byte[]> BLPopBytesWithKey(TimeSpan timeout, params string[] keys)
         {
             return Write(RedisCommands.BLPopBytesWithKey(timeout, keys));
         }
@@ -672,7 +707,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BLPop(timeout, keys));
         }
-		public byte[] BLPopBytes(int timeout, params string[] keys)
+        public byte[] BLPopBytes(int timeout, params string[] keys)
         {
             return Write(RedisCommands.BLPopBytes(timeout, keys));
         }
@@ -687,7 +722,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BLPop(timeout, keys));
         }
-		public byte[] BLPopBytes(TimeSpan timeout, params string[] keys)
+        public byte[] BLPopBytes(TimeSpan timeout, params string[] keys)
         {
             return Write(RedisCommands.BLPopBytes(timeout, keys));
         }
@@ -702,7 +737,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPopWithKey(timeout, keys));
         }
-		public Tuple<string, byte[]> BRPopBytesWithKey(int timeout, params string[] keys)
+        public Tuple<string, byte[]> BRPopBytesWithKey(int timeout, params string[] keys)
         {
             return Write(RedisCommands.BRPopBytesWithKey(timeout, keys));
         }
@@ -717,7 +752,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPopWithKey(timeout, keys));
         }
-		public Tuple<string, byte[]> BRPopBytesWithKey(TimeSpan timeout, params string[] keys)
+        public Tuple<string, byte[]> BRPopBytesWithKey(TimeSpan timeout, params string[] keys)
         {
             return Write(RedisCommands.BRPopBytesWithKey(timeout, keys));
         }
@@ -732,7 +767,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPop(timeout, keys));
         }
-		public byte[] BRPopBytes(int timeout, params string[] keys)
+        public byte[] BRPopBytes(int timeout, params string[] keys)
         {
             return Write(RedisCommands.BRPopBytes(timeout, keys));
         }
@@ -747,7 +782,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPop(timeout, keys));
         }
-		public byte[] BRPopBytes(TimeSpan timeout, params string[] keys)
+        public byte[] BRPopBytes(TimeSpan timeout, params string[] keys)
         {
             return Write(RedisCommands.BRPopBytes(timeout, keys));
         }
@@ -763,7 +798,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPopLPush(source, destination, timeout));
         }
-		public byte[] BRPopBytesLPush(string source, string destination, int timeout)
+        public byte[] BRPopBytesLPush(string source, string destination, int timeout)
         {
             return Write(RedisCommands.BRPopBytesLPush(source, destination, timeout));
         }
@@ -779,7 +814,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.BRPopLPush(source, destination, timeout));
         }
-		public byte[] BRPopBytesLPush(string source, string destination, TimeSpan timeout)
+        public byte[] BRPopBytesLPush(string source, string destination, TimeSpan timeout)
         {
             return Write(RedisCommands.BRPopBytesLPush(source, destination, timeout));
         }
@@ -794,7 +829,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.LIndex(key, index));
         }
-		public byte[] LIndexBytes(string key, long index)
+        public byte[] LIndexBytes(string key, long index)
         {
             return Write(RedisCommands.LIndexBytes(key, index));
         }
@@ -831,7 +866,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.LPop(key));
         }
-		public byte[] LPopBytes(string key)
+        public byte[] LPopBytes(string key)
         {
             return Write(RedisCommands.LPopBytes(key));
         }
@@ -869,7 +904,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.LRange(key, start, stop));
         }
-		public byte[][] LRangeBytes(string key, long start, long stop)
+        public byte[][] LRangeBytes(string key, long start, long stop)
         {
             return Write(RedisCommands.LRangeBytes(key, start, stop));
         }
@@ -919,7 +954,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.RPop(key));
         }
-		public byte[] RPopBytes(string key)
+        public byte[] RPopBytes(string key)
         {
             return Write(RedisCommands.RPopBytes(key));
         }
@@ -934,7 +969,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.RPopLPush(source, destination));
         }
-		public byte[] RPopBytesLPush(string source, string destination)
+        public byte[] RPopBytesLPush(string source, string destination)
         {
             return Write(RedisCommands.RPopBytesLPush(source, destination));
         }
@@ -993,7 +1028,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SDiff(keys));
         }
-		public byte[][] SDiffBytes(params string[] keys)
+        public byte[][] SDiffBytes(params string[] keys)
         {
             return Write(RedisCommands.SDiffBytes(keys));
         }
@@ -1018,7 +1053,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SInter(keys));
         }
-		public byte[][] SInterBytes(params string[] keys)
+        public byte[][] SInterBytes(params string[] keys)
         {
             return Write(RedisCommands.SInterBytes(keys));
         }
@@ -1054,7 +1089,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SMembers(key));
         }
-		public byte[][] SMembersBytes(string key)
+        public byte[][] SMembersBytes(string key)
         {
             return Write(RedisCommands.SMembersBytes(key));
         }
@@ -1080,9 +1115,24 @@ namespace CSRedis
         {
             return Write(RedisCommands.SPop(key));
         }
-		public byte[] SPopBytes(string key)
+        public byte[] SPopBytes(string key)
         {
             return Write(RedisCommands.SPopBytes(key));
+        }
+
+        /// <summary>
+        /// Remove and return one or more random member from a set
+        /// </summary>
+        /// <param name="key">Set key</param>
+        /// <param name="count">Number of elements to remove and return</param>
+        /// <returns>The removed elements</returns>
+        public string[] SPop(string key, long count)
+        {
+            return Write(RedisCommands.SPop(key, count));
+        }
+        public byte[][] SPopBytes(string key, long count)
+        {
+            return Write(RedisCommands.SPopBytes(key, count));
         }
 
         /// <summary>
@@ -1094,7 +1144,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SRandMember(key));
         }
-		public byte[] SRandMemberBytes(string key)
+        public byte[] SRandMemberBytes(string key)
         {
             return Write(RedisCommands.SRandMemberBytes(key));
         }
@@ -1109,7 +1159,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SRandMembers(key, count));
         }
-		public byte[][] SRandMembersBytes(string key, long count)
+        public byte[][] SRandMembersBytes(string key, long count)
         {
             return Write(RedisCommands.SRandMembersBytes(key, count));
         }
@@ -1134,7 +1184,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SUnion(keys));
         }
-		public byte[][] SUnionBytes(params string[] keys)
+        public byte[][] SUnionBytes(params string[] keys)
         {
             return Write(RedisCommands.SUnionBytes(keys));
         }
@@ -1162,13 +1212,30 @@ namespace CSRedis
         {
             return Write(RedisCommands.SScan(key, cursor, pattern, count));
         }
-		public RedisScan<byte[]> SScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        public RedisScan<byte[]> SScanBytes(string key, long cursor, string pattern = null, long? count = null)
         {
             return Write(RedisCommands.SScanBytes(key, cursor, pattern, count));
         }
         #endregion
 
         #region Sorted Sets
+        public Tuple<string, decimal>[] ZPopMax(string key, long count)
+        {
+            return Write(RedisCommands.ZPopMax(key, count));
+        }
+        public Tuple<byte[], decimal>[] ZPopMaxBytes(string key, long count)
+        {
+            return Write(RedisCommands.ZPopMaxBytes(key, count));
+        }
+        public Tuple<string, decimal>[] ZPopMin(string key, long count)
+        {
+            return Write(RedisCommands.ZPopMin(key, count));
+        }
+        public Tuple<byte[], decimal>[] ZPopMinBytes(string key, long count)
+        {
+            return Write(RedisCommands.ZPopMinBytes(key, count));
+        }
+
         /// <summary>
         /// Add one or more members to a sorted set, or update its score if it already exists
         /// </summary>
@@ -1210,7 +1277,7 @@ namespace CSRedis
         /// <param name="exclusiveMin">Minimum score is exclusive</param>
         /// <param name="exclusiveMax">Maximum score is exclusive</param>
         /// <returns>Number of elements in the specified score range</returns>
-        public long ZCount(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)
+        public long ZCount(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             return Write(RedisCommands.ZCount(key, min, max, exclusiveMin, exclusiveMax));
         }
@@ -1234,7 +1301,7 @@ namespace CSRedis
         /// <param name="increment">Increment by value</param>
         /// <param name="member">Sorted set member to increment</param>
         /// <returns>New score of member</returns>
-        public double ZIncrBy(string key, double increment, string member)
+        public decimal ZIncrBy(string key, decimal increment, object member)
         {
             return Write(RedisCommands.ZIncrBy(key, increment, member));
         }
@@ -1247,7 +1314,7 @@ namespace CSRedis
         /// <param name="aggregate">Aggregation function of resulting set</param>
         /// <param name="keys">Sorted set keys to intersect</param>
         /// <returns>Number of elements in the resulting sorted set</returns>
-        public long ZInterStore(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public long ZInterStore(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             return Write(RedisCommands.ZInterStore(destination, weights, aggregate, keys));
         }
@@ -1275,7 +1342,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.ZRange(key, start, stop, withScores));
         }
-		public byte[][] ZRangeBytes(string key, long start, long stop, bool withScores = false)
+        public byte[][] ZRangeBytes(string key, long start, long stop, bool withScores = false)
         {
             return Write(RedisCommands.ZRangeBytes(key, start, stop, withScores));
         }
@@ -1287,11 +1354,11 @@ namespace CSRedis
         /// <param name="start">Start offset</param>
         /// <param name="stop">Stop offset</param>
         /// <returns>Array of elements in the specified range with scores</returns>
-        public Tuple<string, double>[] ZRangeWithScores(string key, long start, long stop)
+        public Tuple<string, decimal>[] ZRangeWithScores(string key, long start, long stop)
         {
             return Write(RedisCommands.ZRangeWithScores(key, start, stop));
         }
-		public Tuple<byte[], double>[] ZRangeBytesWithScores(string key, long start, long stop)
+        public Tuple<byte[], decimal>[] ZRangeBytesWithScores(string key, long start, long stop)
         {
             return Write(RedisCommands.ZRangeBytesWithScores(key, start, stop));
         }
@@ -1308,11 +1375,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public string[] ZRangeByScore(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public string[] ZRangeByScore(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeByScore(key, min, max, withScores, exclusiveMin, exclusiveMax, offset, count));
         }
-		public byte[][] ZRangeBytesByScore(string key, double min, double max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public byte[][] ZRangeBytesByScore(string key, decimal min, decimal max, bool withScores = false, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeBytesByScore(key, min, max, withScores, exclusiveMin, exclusiveMax, offset, count));
         }
@@ -1331,7 +1398,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.ZRangeByScore(key, min, max, withScores, offset, count));
         }
-		public byte[][] ZRangeBytesByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
+        public byte[][] ZRangeBytesByScore(string key, string min, string max, bool withScores = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeBytesByScore(key, min, max, withScores, offset, count));
         }
@@ -1347,11 +1414,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Tuple<string, double>[] ZRangeByScoreWithScores(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Tuple<string, decimal>[] ZRangeByScoreWithScores(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeByScoreWithScores(key, min, max, exclusiveMin, exclusiveMax, offset, count));
         }
-		public Tuple<byte[], double>[] ZRangeBytesByScoreWithScores(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
+        public Tuple<byte[], decimal>[] ZRangeBytesByScoreWithScores(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeBytesByScoreWithScores(key, min, max, exclusiveMin, exclusiveMax, offset, count));
         }
@@ -1365,11 +1432,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Tuple<string, double>[] ZRangeByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
+        public Tuple<string, decimal>[] ZRangeByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeByScoreWithScores(key, min, max, offset, count));
         }
-		public Tuple<byte[], double>[] ZRangeBytesByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
+        public Tuple<byte[], decimal>[] ZRangeBytesByScoreWithScores(string key, string min, string max, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeBytesByScoreWithScores(key, min, max, offset, count));
         }
@@ -1417,11 +1484,11 @@ namespace CSRedis
         /// <param name="exclusiveMin">Minimum score is exclusive</param>
         /// <param name="exclusiveMax">Maximum score is exclusive</param>
         /// <returns>Number of elements removed</returns>
-        public long ZRemRangeByScore(string key, double min, double max, bool exclusiveMin = false, bool exclusiveMax = false)
+        public long ZRemRangeByScore(string key, decimal min, decimal max, bool exclusiveMin = false, bool exclusiveMax = false)
         {
             return Write(RedisCommands.ZRemRangeByScore(key, min, max, exclusiveMin, exclusiveMax));
         }
-		public long ZRemRangeByScore(string key, string min, string max)
+        public long ZRemRangeByScore(string key, string min, string max)
         {
             return Write(RedisCommands.ZRemRangeByScore(key, min, max));
         }
@@ -1438,7 +1505,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.ZRevRange(key, start, stop, withScores));
         }
-		public byte[][] ZRevRangeBytes(string key, long start, long stop, bool withScores = false)
+        public byte[][] ZRevRangeBytes(string key, long start, long stop, bool withScores = false)
         {
             return Write(RedisCommands.ZRevRangeBytes(key, start, stop, withScores));
         }
@@ -1450,11 +1517,11 @@ namespace CSRedis
         /// <param name="start">Start offset</param>
         /// <param name="stop">Stop offset</param>
         /// <returns>List of elements in the specified range (with optional scores)</returns>
-        public Tuple<string, double>[] ZRevRangeWithScores(string key, long start, long stop)
+        public Tuple<string, decimal>[] ZRevRangeWithScores(string key, long start, long stop)
         {
             return Write(RedisCommands.ZRevRangeWithScores(key, start, stop));
         }
-		public Tuple<byte[], double>[] ZRevRangeBytesWithScores(string key, long start, long stop)
+        public Tuple<byte[], decimal>[] ZRevRangeBytesWithScores(string key, long start, long stop)
         {
             return Write(RedisCommands.ZRevRangeBytesWithScores(key, start, stop));
         }
@@ -1471,11 +1538,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public string[] ZRevRangeByScore(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public string[] ZRevRangeByScore(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeByScore(key, max, min, withScores, exclusiveMax, exclusiveMin, offset, count));
         }
-		public byte[][] ZRevRangeBytesByScore(string key, double max, double min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public byte[][] ZRevRangeBytesByScore(string key, decimal max, decimal min, bool withScores = false, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeBytesByScore(key, max, min, withScores, exclusiveMax, exclusiveMin, offset, count));
         }
@@ -1494,7 +1561,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.ZRevRangeByScore(key, max, min, withScores, offset, count));
         }
-		public byte[][] ZRevRangeBytesByScore(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
+        public byte[][] ZRevRangeBytesByScore(string key, string max, string min, bool withScores = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeBytesByScore(key, max, min, withScores, offset, count));
         }
@@ -1510,11 +1577,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public Tuple<string, double>[] ZRevRangeByScoreWithScores(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public Tuple<string, decimal>[] ZRevRangeByScoreWithScores(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeByScoreWithScores(key, max, min, exclusiveMax, exclusiveMin, offset, count));
         }
-		public Tuple<byte[], double>[] ZRevRangeBytesByScoreWithScores(string key, double max, double min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
+        public Tuple<byte[], decimal>[] ZRevRangeBytesByScoreWithScores(string key, decimal max, decimal min, bool exclusiveMax = false, bool exclusiveMin = false, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeBytesByScoreWithScores(key, max, min, exclusiveMax, exclusiveMin, offset, count));
         }
@@ -1528,11 +1595,11 @@ namespace CSRedis
         /// <param name="offset">Start offset</param>
         /// <param name="count">Number of elements to return</param>
         /// <returns>List of elements in the specified score range (with optional scores)</returns>
-        public Tuple<string, double>[] ZRevRangeByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
+        public Tuple<string, decimal>[] ZRevRangeByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeByScoreWithScores(key, max, min, offset, count));
         }
-		public Tuple<byte[], double>[] ZRevRangeBytesByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
+        public Tuple<byte[], decimal>[] ZRevRangeBytesByScoreWithScores(string key, string max, string min, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRevRangeBytesByScoreWithScores(key, max, min, offset, count));
         }
@@ -1554,7 +1621,7 @@ namespace CSRedis
         /// <param name="key">Sorted set key</param>
         /// <param name="member">Member to lookup</param>
         /// <returns>Score of member, or null if member does not exist</returns>
-        public double? ZScore(string key, object member)
+        public decimal? ZScore(string key, object member)
         {
             return Write(RedisCommands.ZScore(key, member));
         }
@@ -1567,7 +1634,7 @@ namespace CSRedis
         /// <param name="aggregate">Aggregation function of resulting set</param>
         /// <param name="keys">Sorted set keys to union</param>
         /// <returns>Number of elements in the resulting sorted set</returns>
-        public long ZUnionStore(string destination, double[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
+        public long ZUnionStore(string destination, decimal[] weights = null, RedisAggregate? aggregate = null, params string[] keys)
         {
             return Write(RedisCommands.ZUnionStore(destination, weights, aggregate, keys));
         }
@@ -1591,11 +1658,11 @@ namespace CSRedis
         /// <param name="pattern">Glob-style pattern to filter returned elements</param>
         /// <param name="count">Maximum number of elements to return</param>
         /// <returns>Updated cursor and result set</returns>
-        public RedisScan<Tuple<string, double>> ZScan(string key, long cursor, string pattern = null, long? count = null)
+        public RedisScan<Tuple<string, decimal>> ZScan(string key, long cursor, string pattern = null, long? count = null)
         {
             return Write(RedisCommands.ZScan(key, cursor, pattern, count));
         }
-		public RedisScan<Tuple<byte[], double>> ZScanBytes(string key, long cursor, string pattern = null, long? count = null)
+        public RedisScan<Tuple<byte[], decimal>> ZScanBytes(string key, long cursor, string pattern = null, long? count = null)
         {
             return Write(RedisCommands.ZScanBytes(key, cursor, pattern, count));
         }
@@ -1613,7 +1680,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.ZRangeByLex(key, min, max, offset, count));
         }
-		public byte[][] ZRangeBytesByLex(string key, string min, string max, long? offset = null, long? count = null)
+        public byte[][] ZRangeBytesByLex(string key, string min, string max, long? offset = null, long? count = null)
         {
             return Write(RedisCommands.ZRangeBytesByLex(key, min, max, offset, count));
         }
@@ -1747,12 +1814,12 @@ namespace CSRedis
             return Write(RedisCommands.EvalSHA(sha1, keys, arguments));
         }
 
-		/// <summary>
-		/// Check existence of script SHA hashes in the script cache
-		/// </summary>
-		/// <param name="sha1s">SHA1 script hashes</param>
-		/// <returns>Array of boolean values indicating script existence on server</returns>
-		public bool[] ScriptExists(params string[] sha1s)
+        /// <summary>
+        /// Check existence of script SHA hashes in the script cache
+        /// </summary>
+        /// <param name="sha1s">SHA1 script hashes</param>
+        /// <returns>Array of boolean values indicating script existence on server</returns>
+        public bool[] ScriptExists(params string[] sha1s)
         {
             return Write(RedisCommands.ScriptExists(sha1s));
         }
@@ -1865,18 +1932,18 @@ namespace CSRedis
         {
             return Write(RedisCommands.Get(key));
         }
-		public byte[] GetBytes(string key)
-		{
-			return Write(RedisCommands.GetBytes(key));
-		}
+        public byte[] GetBytes(string key)
+        {
+            return Write(RedisCommands.GetBytes(key));
+        }
 
-		/// <summary>
-		/// Returns the bit value at offset in the string value stored at key
-		/// </summary>
-		/// <param name="key">Key to lookup</param>
-		/// <param name="offset">Offset of key to check</param>
-		/// <returns>Bit value stored at offset</returns>
-		public bool GetBit(string key, uint offset)
+        /// <summary>
+        /// Returns the bit value at offset in the string value stored at key
+        /// </summary>
+        /// <param name="key">Key to lookup</param>
+        /// <param name="offset">Offset of key to check</param>
+        /// <returns>Bit value stored at offset</returns>
+        public bool GetBit(string key, uint offset)
         {
             return Write(RedisCommands.GetBit(key, offset));
         }
@@ -1892,7 +1959,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.GetRange(key, start, end));
         }
-		public byte[] GetRangeBytes(string key, long start, long end)
+        public byte[] GetRangeBytes(string key, long start, long end)
         {
             return Write(RedisCommands.GetRangeBytes(key, start, end));
         }
@@ -1907,7 +1974,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.GetSet(key, value));
         }
-		public byte[] GetSetBytes(string key, object value)
+        public byte[] GetSetBytes(string key, object value)
         {
             return Write(RedisCommands.GetSetBytes(key, value));
         }
@@ -1939,7 +2006,7 @@ namespace CSRedis
         /// <param name="key">Key to modify</param>
         /// <param name="increment">Increment amount</param>
         /// <returns>Value of key after increment</returns>
-        public double IncrByFloat(string key, double increment)
+        public decimal IncrByFloat(string key, decimal increment)
         {
             return Write(RedisCommands.IncrByFloat(key, increment));
         }
@@ -1953,7 +2020,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.MGet(keys));
         }
-		public byte[][] MGetBytes(params string[] keys)
+        public byte[][] MGetBytes(params string[] keys)
         {
             return Write(RedisCommands.MGetBytes(keys));
         }
@@ -2358,7 +2425,7 @@ namespace CSRedis
         {
             return Write(RedisCommands.SlaveOfNoOne());
         }
-        
+
         /// <summary>
         /// Get latest entries from the slow log
         /// </summary>
@@ -2489,99 +2556,149 @@ namespace CSRedis
         {
             return Write(RedisCommands.PfMerge(destKey, sourceKeys));
         }
-		#endregion
+        #endregion
 
-		#region Geo redis-server 3.2
-		public long GeoAdd(string key, params (double longitude, double latitude, object member)[] values) {
-			if (values == null || values.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			foreach (var v in values) args.AddRange(new object[] { v.longitude, v.latitude, v.member });
-			return Write(new RedisInt("GEOADD", args.ToArray()));
-		}
-		public double? GeoDist(string key, object member1, object member2, GeoUnit unit = GeoUnit.m) {
-			if (unit ==  GeoUnit.m) return Write(new RedisFloat.Nullable("GEODIST", key, member1, member2));
-			return Write(new RedisFloat.Nullable("GEODIST", key, member1, member2, unit));
-		}
-		public string[] GeoHash(string key, object[] members) {
-			if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			args.AddRange(members);
-			return Write(new RedisArray.Strings("GEOHASH", args.ToArray()));
-		}
-		public (double longitude, double latitude)?[] GeoPos(string key, object[] members) {
-			if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
-			var args = new List<object>();
-			args.Add(key);
-			args.AddRange(members);
-			var ret = Write(new RedisArray.Generic<double[]>(new RedisArray.Generic<double>(new RedisFloat("GEOPOS", args.ToArray()))));
-			return ret.Select(a => a != null && a.Length == 2 ? new(double, double)?((a[0], a[1])) : null).ToArray();
-		}
-		public (string member, double dist, double longitude, double latitude, long hash)[] GeoRadius(string key, double longitude, double latitude, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+        #region Geo redis-server 3.2
+        public long GeoAdd(string key, params (decimal longitude, decimal latitude, object member)[] values)
+        {
+            if (values == null || values.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            foreach (var v in values) args.AddRange(new object[] { v.longitude, v.latitude, v.member });
+            return Write(new RedisInt("GEOADD", args.ToArray()));
+        }
+        public decimal? GeoDist(string key, object member1, object member2, GeoUnit unit = GeoUnit.m)
+        {
+            if (unit == GeoUnit.m) return Write(new RedisFloat.Nullable("GEODIST", key, member1, member2));
+            return Write(new RedisFloat.Nullable("GEODIST", key, member1, member2, unit));
+        }
+        public string[] GeoHash(string key, object[] members)
+        {
+            if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(members);
+            return Write(new RedisArray.Strings("GEOHASH", args.ToArray()));
+        }
+        public (decimal longitude, decimal latitude)?[] GeoPos(string key, object[] members)
+        {
+            if (members == null || members.Length == 0) throw new Exception("values 参数不能为空");
+            var args = new List<object>();
+            args.Add(key);
+            args.AddRange(members);
+            var ret = Write(new RedisArray.Generic<decimal[]>(new RedisArray.Generic<decimal>(new RedisFloat("GEOPOS", args.ToArray()))));
+            return ret.Select(a => a != null && a.Length == 2 ? new (decimal, decimal)?((a[0], a[1])) : null).ToArray();
+        }
+        public (string member, decimal dist, decimal longitude, decimal latitude, long hash)[] GeoRadius(string key, decimal longitude, decimal latitude, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.AddRange(new object[] { "COUNT", count });
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<string, double, long, double[]>.Single(
-				new RedisString(null), 
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null :  new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
-			var ret = Write(new RedisArray.Generic<Tuple<string, double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		public (byte[] member, double dist, double longitude, double latitude, long hash)[] GeoRadiusBytes(string key, double longitude, double latitude, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<string, decimal, long, decimal[]>.Single(
+                new RedisString(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
+            var ret = Write(new RedisArray.Generic<Tuple<string, decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        public (byte[] member, decimal dist, decimal longitude, decimal latitude, long hash)[] GeoRadiusBytes(string key, decimal longitude, decimal latitude, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, longitude, latitude, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.AddRange(new object[] { "COUNT", count });
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<byte[], double, long, double[]>.Single(
-				new RedisBytes(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
-			var ret = Write(new RedisArray.Generic<Tuple<byte[], double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		public (string member, double dist, double longitude, double latitude, long hash)[] GeoRadiusByMember(string key, object member, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, member, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<byte[], decimal, long, decimal[]>.Single(
+                new RedisBytes(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUS", args.ToArray());
+            var ret = Write(new RedisArray.Generic<Tuple<byte[], decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        public (string member, decimal dist, decimal longitude, decimal latitude, long hash)[] GeoRadiusByMember(string key, object member, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, member, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.AddRange(new object[] { "COUNT", count });
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<string, double, long, double[]>.Single(
-				new RedisString(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
-			var ret = Write(new RedisArray.Generic<Tuple<string, double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		public (byte[] member, double dist, double longitude, double latitude, long hash)[] GeoRadiusBytesByMember(string key, object member, double radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false) {
-			var args = new List<object>(new object[] { key, member, radius, unit });
-			if (withCoord) args.Add("WITHCOORD");
-			if (withDist) args.Add("WITHDIST");
-			if (withHash) args.Add("WITHHASH");
-			if (count.HasValue) args.Add(count);
-			if (sorting.HasValue) args.Add(sorting);
+            var cmd = new RedisTuple.Generic<string, decimal, long, decimal[]>.Single(
+                new RedisString(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
+            var ret = Write(new RedisArray.Generic<Tuple<string, decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        public (byte[] member, decimal dist, decimal longitude, decimal latitude, long hash)[] GeoRadiusBytesByMember(string key, object member, decimal radius, GeoUnit unit = GeoUnit.m, long? count = null, GeoOrderBy? sorting = null, bool withCoord = false, bool withDist = false, bool withHash = false)
+        {
+            var args = new List<object>(new object[] { key, member, radius, unit });
+            if (withCoord) args.Add("WITHCOORD");
+            if (withDist) args.Add("WITHDIST");
+            if (withHash) args.Add("WITHHASH");
+            if (count.HasValue) args.AddRange(new object[] { "COUNT", count });
+            if (sorting.HasValue) args.Add(sorting);
 
-			var cmd = new RedisTuple.Generic<byte[], double, long, double[]>.Single(
-				new RedisBytes(null),
-				withDist == false ? null : new RedisFloat(null),
-				withHash == false ? null : new RedisInt(null),
-				withCoord == false ? null : new RedisArray.Generic<double>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
-			var ret = Write(new RedisArray.Generic<Tuple<byte[], double, long, double[]>>(cmd));
-			return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(double) : a.Item4[0], a.Item4 == null ? default(double) : a.Item4[1], a.Item3)).ToArray();
-		}
-		#endregion
-	}
+            var cmd = new RedisTuple.Generic<byte[], decimal, long, decimal[]>.Single(
+                new RedisBytes(null),
+                withDist == false ? null : new RedisFloat(null),
+                withHash == false ? null : new RedisInt(null),
+                withCoord == false ? null : new RedisArray.Generic<decimal>(new RedisFloat(null)), "GEORADIUSBYMEMBER", args.ToArray());
+            var ret = Write(new RedisArray.Generic<Tuple<byte[], decimal, long, decimal[]>>(cmd));
+            return ret.Select(a => (a.Item1, a.Item2, a.Item4 == null ? default(decimal) : a.Item4[0], a.Item4 == null ? default(decimal) : a.Item4[1], a.Item3)).ToArray();
+        }
+        #endregion
+
+        #region Streams 5.0
+        public long XAck(string key, string group, params string[] id) => Write(RedisCommands.XAck(key, group, id));
+
+        public string XAdd(string key, long maxLen, string id = "*", params (string, string)[] fieldValues) => Write(RedisCommands.XAdd(key, maxLen, id, fieldValues));
+
+        public (string id, string[] items)[] XClaim(string key, string group, string consumer, long minIdleTime, params string[] id) =>
+            Write(RedisCommands.XClaim(key, group, consumer, minIdleTime, id));
+        public (string id, string[] items)[] XClaim(string key, string group, string consumer, long minIdleTime, string[] id, long idle, long retryCount, bool force) =>
+            Write(RedisCommands.XClaim(key, group, consumer, minIdleTime, id, idle, retryCount, force));
+
+        public string[] XClaimJustId(string key, string group, string consumer, long minIdleTime, params string[] id) =>
+            Write(RedisCommands.XClaimJustId(key, group, consumer, minIdleTime, id));
+        public string[] XClaimJustId(string key, string group, string consumer, long minIdleTime, string[] id, long idle, long retryCount, bool force) =>
+            Write(RedisCommands.XClaimJustId(key, group, consumer, minIdleTime, id, idle, retryCount, force));
+
+        public long XDel(string key, params string[] id) => Write(RedisCommands.XDel(key, id));
+
+        public string XGroupCreate(string key, string group, string id = "$", bool MkStream = false) => Write(RedisCommands.XGroup.Create(key, group, id, MkStream));
+        public string XGroupSetId(string key, string group, string id = "$") => Write(RedisCommands.XGroup.SetId(key, group, id));
+        public bool XGroupDestroy(string key, string group) => Write(RedisCommands.XGroup.Destroy(key, group));
+        public bool XGroupDelConsumer(string key, string group, string consumer) => Write(RedisCommands.XGroup.DelConsumer(key, group, consumer));
+
+        public (long length, long radixTreeKeys, long radixTreeNodes, long groups, string lastGeneratedId, (string id, string[] items) firstEntry, (string id, string[] items) lastEntry) XInfoStream(string key) => Write(RedisCommands.XInfoStream(key));
+        public (string name, long consumers, long pending, string lastDeliveredId)[] XInfoGroups(string key) => Write(RedisCommands.XInfoGroups(key));
+        public (string name, long pending, long idle)[] XInfoConsumers(string key, string group) => Write(RedisCommands.XInfoConsumers(key, group));
+
+        public long XLen(string key) => Write(RedisCommands.XLen(key));
+
+        public (long count, string minId, string maxId, (string consumer, long count)[] pendings) XPending(string key, string group) => Write(RedisCommands.XPending(key, group));
+        public (string id, string consumer, long millisecond, long transferTimes)[] XPending(string key, string group, string start, string end, long count, string consumer = null) => Write(RedisCommands.XPending(key, group, start, end, count, consumer));
+
+        public (string id, string[] items)[] XRange(string key, string start, string end, long count = 1) => Write(RedisCommands.XRange(key, start, end, count));
+        public (string id, string[] items)[] XRevRange(string key, string end, string start, long count = 1) => Write(RedisCommands.XRevRange(key, end, start, count));
+
+        public (string key, (string id, string[] items)[] data)[] XRead(long count, long block, params (string key, string id)[] streams) =>
+            Write(RedisCommands.XRead(count, block, streams));
+        public (string key, (string id, string[] items)[] data)[] XReadGroup(string group, string consumer, long count, long block, params (string key, string id)[] streams) =>
+            Write(RedisCommands.XReadGroup(group, consumer, count, block, streams));
+
+        public long XTrim(string key, long maxLen) => Write(RedisCommands.XTrim(key, maxLen));
+        #endregion
+    }
 }
