@@ -9,7 +9,7 @@ namespace CSRedisCore.Tests
     {
 		//测试 redis-cluster 不能设置 defaultDatabase
 
-		protected CSRedisClient rds = new CSRedisClient("127.0.0.1:6371,defaultDatabase=0,poolsize=3");
+		protected CSRedisClient rds = new CSRedisClient("127.0.0.1,defaultDatabase=0,poolsize=3,tryit=0");
 
 		protected readonly object Null = null;
 		protected readonly string String = "我是中国人";
@@ -17,7 +17,7 @@ namespace CSRedisCore.Tests
 		protected readonly TestClass Class = new TestClass { Id = 1, Name = "Class名称", CreateTime = DateTime.Now, TagId = new[] { 1, 3, 3, 3, 3 } };
 
 		public TestBase() {
-			rds.NodesServerManager.FlushAll();
+			//rds.NodesServerManager.FlushAll();
 		}
     }
 
